% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{adjustments}
\alias{adjustments}
\title{adjustments}
\usage{
outline$adjustments()

solid$adjustments()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''adjustments'

Returns the svg markup for the heroicon ''adjustments'
}
\description{
adjustments

adjustments
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/adjustments.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/adjustments.svg}
}
\keyword{adjustments}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
