% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{arrow_narrow_down}
\alias{arrow_narrow_down}
\title{arrow_narrow_down}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''arrow_narrow_down'

Returns the svg markup for the heroicon ''arrow_narrow_down'
}
\description{
arrow_narrow_down

arrow_narrow_down
}
\examples{
rheroicons::outline$arrow_narrow_down(
  id = 'my_arrow_narrow_down_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the arrow_narrow_down icon'
)
rheroicons::solid$arrow_narrow_down(
  id = 'my_arrow_narrow_down_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the arrow_narrow_down icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/arrow_narrow_down.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/arrow_narrow_down.svg}
}
\keyword{arrow_narrow_down}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
