% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{at_symbol}
\alias{at_symbol}
\title{at_symbol}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a at_symbol
}
\description{
Render an svg icon of a at_symbol
}
\examples{
rheroicons::icons$at_symbol(
  type = "solid",
  id = "my_at_symbol_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the at_symbol icon"
)
}
\keyword{at_symbol}
\keyword{rheroicons}
