% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{badge_check}
\alias{badge_check}
\title{badge_check}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a badge_check
}
\description{
Render an svg icon of a badge_check
}
\examples{
rheroicons::icons$badge_check(
  type = "solid",
  id = "my_badge_check_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the badge_check icon"
)
}
\keyword{badge_check}
\keyword{rheroicons}
