% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{ban}
\alias{ban}
\title{ban}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''ban'

Returns the svg markup for the heroicon ''ban'
}
\description{
ban

ban
}
\examples{
rheroicons::outline$ban(
  id = 'my_ban_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the ban icon'
)
rheroicons::solid$ban(
  id = 'my_ban_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the ban icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/ban.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/ban.svg}
}
\keyword{ban}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
