% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{bell}
\alias{bell}
\title{bell}
\usage{
outline$bell()

solid$bell()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a bell

create a solid or outline SVG icon of a bell
}
\description{
bell

bell
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/bell.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/bell.svg}
\url{https://heroicons.dev}
}
\keyword{bell}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
