% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{bookmark}
\alias{bookmark}
\title{bookmark}
\usage{
outline$bookmark()

solid$bookmark()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a bookmark

create a solid or outline SVG icon of a bookmark
}
\description{
bookmark

bookmark
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/bookmark.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/bookmark.svg}
\url{https://heroicons.dev}
}
\keyword{bookmark}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
