% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{briefcase}
\alias{briefcase}
\title{briefcase}
\usage{
outline$briefcase()

solid$briefcase()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a briefcase

create a solid or outline SVG icon of a briefcase
}
\description{
briefcase

briefcase
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/briefcase.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/briefcase.svg}
\url{https://heroicons.dev}
}
\keyword{briefcase}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
