% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{chart_bar}
\alias{chart_bar}
\title{chart_bar}
\usage{
outline$chart_bar()

solid$chart_bar()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a chart_bar

create a solid or outline SVG icon of a chart_bar
}
\description{
chart_bar

chart_bar
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/chart_bar.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/chart_bar.svg}
\url{https://heroicons.dev}
}
\keyword{chart_bar}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
