% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{chart_pie}
\alias{chart_pie}
\title{chart_pie}
\usage{
outline$chart_pie()

solid$chart_pie()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a chart_pie

create a solid or outline SVG icon of a chart_pie
}
\description{
chart_pie

chart_pie
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/chart_pie.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/chart_pie.svg}
\url{https://heroicons.dev}
}
\keyword{chart_pie}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
