% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{chat}
\alias{chat}
\title{chat}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''chat'

Returns the svg markup for the heroicon ''chat'
}
\description{
chat

chat
}
\examples{
rheroicons::outline$chat(
  id = 'my_chat_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the chat icon'
)
rheroicons::solid$chat(
  id = 'my_chat_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the chat icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/chat.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/chat.svg}
}
\keyword{chat}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
