% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{chevron_double_right}
\alias{chevron_double_right}
\title{chevron_double_right}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a chevron_double_right
}
\description{
Render an svg icon of a chevron_double_right
}
\examples{
rheroicons::icons$chevron_double_right(
  type = "solid",
  id = "my_chevron_double_right_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the chevron_double_right icon"
)
}
\keyword{chevron_double_right}
\keyword{rheroicons}
