% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{chevron_right}
\alias{chevron_right}
\title{chevron_right}
\usage{
outline$chevron_right()

solid$chevron_right()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a chevron_right

create a solid or outline SVG icon of a chevron_right
}
\description{
chevron_right

chevron_right
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/chevron_right.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/chevron_right.svg}
\url{https://heroicons.dev}
}
\keyword{chevron_right}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
