% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{clipboard_check}
\alias{clipboard_check}
\title{clipboard_check}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''clipboard_check'

Returns the svg markup for the heroicon ''clipboard_check'
}
\description{
clipboard_check

clipboard_check
}
\examples{
rheroicons::outline$clipboard_check(
  id = 'my_clipboard_check_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the clipboard_check icon'
)
rheroicons::solid$clipboard_check(
  id = 'my_clipboard_check_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the clipboard_check icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/clipboard_check.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/clipboard_check.svg}
}
\keyword{clipboard_check}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
