% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{clipboard_copy}
\alias{clipboard_copy}
\title{clipboard_copy}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''clipboard_copy'

Returns the svg markup for the heroicon ''clipboard_copy'
}
\description{
clipboard_copy

clipboard_copy
}
\examples{
rheroicons::outline$clipboard_copy(
  id = 'my_clipboard_copy_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the clipboard_copy icon'
)
rheroicons::solid$clipboard_copy(
  id = 'my_clipboard_copy_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the clipboard_copy icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/clipboard_copy.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/clipboard_copy.svg}
}
\keyword{clipboard_copy}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
