% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{clipboard_list}
\alias{clipboard_list}
\title{clipboard_list}
\usage{
outline$clipboard_list()

solid$clipboard_list()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a clipboard_list

create a solid or outline SVG icon of a clipboard_list
}
\description{
clipboard_list

clipboard_list
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/clipboard_list.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/clipboard_list.svg}
\url{https://heroicons.dev}
}
\keyword{clipboard_list}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
