% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{clock}
\alias{clock}
\title{clock}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''clock'

Returns the svg markup for the heroicon ''clock'
}
\description{
clock

clock
}
\examples{
rheroicons::outline$clock(
  id = 'my_clock_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the clock icon'
)
rheroicons::solid$clock(
  id = 'my_clock_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the clock icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/clock.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/clock.svg}
}
\keyword{clock}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
