% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{cloud_download}
\alias{cloud_download}
\title{cloud_download}
\usage{
outline$cloud_download()

solid$cloud_download()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''cloud_download'

Returns the svg markup for the heroicon ''cloud_download'
}
\description{
cloud_download

cloud_download
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/cloud_download.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/cloud_download.svg}
}
\keyword{cloud_download}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
