% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{cog}
\alias{cog}
\title{cog}
\usage{
outline$cog()

solid$cog()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a cog

create a solid or outline SVG icon of a cog
}
\description{
cog

cog
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/cog.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/cog.svg}
\url{https://heroicons.dev}
}
\keyword{cog}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
