% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{collection}
\alias{collection}
\title{collection}
\usage{
outline$collection()

solid$collection()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''collection'

Returns the svg markup for the heroicon ''collection'
}
\description{
collection

collection
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/collection.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/collection.svg}
}
\keyword{collection}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
