% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{cube_transparent}
\alias{cube_transparent}
\title{cube_transparent}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a cube_transparent
}
\description{
Render an svg icon of a cube_transparent
}
\examples{
rheroicons::icons$cube_transparent(
  type = "solid",
  id = "my_cube_transparent_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the cube_transparent icon"
)
}
\keyword{cube_transparent}
\keyword{rheroicons}
