% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{currency_dollar}
\alias{currency_dollar}
\title{currency_dollar}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''currency_dollar'

Returns the svg markup for the heroicon ''currency_dollar'
}
\description{
currency_dollar

currency_dollar
}
\examples{
rheroicons::outline$currency_dollar(
  id = 'my_currency_dollar_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the currency_dollar icon'
)
rheroicons::solid$currency_dollar(
  id = 'my_currency_dollar_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the currency_dollar icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/currency_dollar.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/currency_dollar.svg}
}
\keyword{currency_dollar}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
