% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{currency_euro}
\alias{currency_euro}
\title{currency_euro}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''currency_euro'

Returns the svg markup for the heroicon ''currency_euro'
}
\description{
currency_euro

currency_euro
}
\examples{
rheroicons::outline$currency_euro(
  id = 'my_currency_euro_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the currency_euro icon'
)
rheroicons::solid$currency_euro(
  id = 'my_currency_euro_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the currency_euro icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/currency_euro.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/currency_euro.svg}
}
\keyword{currency_euro}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
