% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{currency_yen}
\alias{currency_yen}
\title{currency_yen}
\usage{
outline$currency_yen()

solid$currency_yen()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a currency_yen

create a solid or outline SVG icon of a currency_yen
}
\description{
currency_yen

currency_yen
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/currency_yen.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/currency_yen.svg}
\url{https://heroicons.dev}
}
\keyword{currency_yen}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
