% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{desktop_computer}
\alias{desktop_computer}
\title{desktop_computer}
\usage{
outline$desktop_computer()

solid$desktop_computer()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''desktop_computer'

Returns the svg markup for the heroicon ''desktop_computer'
}
\description{
desktop_computer

desktop_computer
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/desktop_computer.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/desktop_computer.svg}
}
\keyword{desktop_computer}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
