% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{device_mobile}
\alias{device_mobile}
\title{device_mobile}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a device_mobile
}
\description{
Render an svg icon of a device_mobile
}
\examples{
rheroicons::icons$device_mobile(
  type = "solid",
  id = "my_device_mobile_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the device_mobile icon"
)
}
\keyword{device_mobile}
\keyword{rheroicons}
