% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{document_add}
\alias{document_add}
\title{document_add}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''document_add'

Returns the svg markup for the heroicon ''document_add'
}
\description{
document_add

document_add
}
\examples{
rheroicons::outline$document_add(
  id = 'my_document_add_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the document_add icon'
)
rheroicons::solid$document_add(
  id = 'my_document_add_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the document_add icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/document_add.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/document_add.svg}
}
\keyword{document_add}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
