% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{document_remove}
\alias{document_remove}
\title{document_remove}
\usage{
outline$document_remove()

solid$document_remove()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a document_remove

create a solid or outline SVG icon of a document_remove
}
\description{
document_remove

document_remove
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/document_remove.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/document_remove.svg}
\url{https://heroicons.dev}
}
\keyword{document_remove}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
