% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{dots_horizontal}
\alias{dots_horizontal}
\title{dots_horizontal}
\usage{
outline$dots_horizontal()

solid$dots_horizontal()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a dots_horizontal

create a solid or outline SVG icon of a dots_horizontal
}
\description{
dots_horizontal

dots_horizontal
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/dots_horizontal.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/dots_horizontal.svg}
\url{https://heroicons.dev}
}
\keyword{dots_horizontal}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
