% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{download}
\alias{download}
\title{download}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''download'

Returns the svg markup for the heroicon ''download'
}
\description{
download

download
}
\examples{
rheroicons::outline$download(
  id = 'my_download_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the download icon'
)
rheroicons::solid$download(
  id = 'my_download_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the download icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/download.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/download.svg}
}
\keyword{download}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
