% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{emoji_sad}
\alias{emoji_sad}
\title{emoji_sad}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''emoji_sad'

Returns the svg markup for the heroicon ''emoji_sad'
}
\description{
emoji_sad

emoji_sad
}
\examples{
rheroicons::outline$emoji_sad(
  id = 'my_emoji_sad_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the emoji_sad icon'
)
rheroicons::solid$emoji_sad(
  id = 'my_emoji_sad_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the emoji_sad icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/emoji_sad.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/emoji_sad.svg}
}
\keyword{emoji_sad}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
