% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{exclamation}
\alias{exclamation}
\title{exclamation}
\usage{
outline$exclamation()

solid$exclamation()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''exclamation'

Returns the svg markup for the heroicon ''exclamation'
}
\description{
exclamation

exclamation
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/exclamation.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/exclamation.svg}
}
\keyword{exclamation}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
