% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{filter}
\alias{filter}
\title{filter}
\usage{
outline$filter()

solid$filter()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''filter'

Returns the svg markup for the heroicon ''filter'
}
\description{
filter

filter
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/filter.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/filter.svg}
}
\keyword{filter}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
