% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_icon.R
\name{find_icons}
\alias{find_icons}
\title{Find Icons}
\usage{
find_icons(query = "")
}
\arguments{
\item{query}{a search term passed down to \code{stringr::str_subset}}
}
\value{
Find an icon using a search term
}
\description{
Find an icon using a search term
}
\examples{
find_icons(query = "chevron")
find_icons(query = "chevron|arrow")
find_icons(query = "down|up|left|right")

}
