% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{finger_print}
\alias{finger_print}
\title{finger_print}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a finger_print
}
\description{
Render an svg icon of a finger_print
}
\examples{
rheroicons::icons$finger_print(
  type = "solid",
  id = "my_finger_print_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the finger_print icon"
)
}
\keyword{finger_print}
\keyword{rheroicons}
