% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{folder_download}
\alias{folder_download}
\title{folder_download}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a folder_download
}
\description{
Render an svg icon of a folder_download
}
\examples{
rheroicons::icons$folder_download(
  type = "solid",
  id = "my_folder_download_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the folder_download icon"
)
}
\keyword{folder_download}
\keyword{rheroicons}
