% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{gift}
\alias{gift}
\title{gift}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a gift
}
\description{
Render an svg icon of a gift
}
\examples{
rheroicons::icons$gift(
  type = "solid",
  id = "my_gift_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the gift icon"
)
}
\keyword{gift}
\keyword{rheroicons}
