% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{globe}
\alias{globe}
\title{globe}
\usage{
outline$globe()

solid$globe()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a globe

create a solid or outline SVG icon of a globe
}
\description{
globe

globe
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/globe.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/globe.svg}
\url{https://heroicons.dev}
}
\keyword{globe}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
