% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{hand}
\alias{hand}
\title{hand}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a hand
}
\description{
Render an svg icon of a hand
}
\examples{
rheroicons::icons$hand(
  type = "solid",
  id = "my_hand_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the hand icon"
)
}
\keyword{hand}
\keyword{rheroicons}
