% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{heart}
\alias{heart}
\title{heart}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''heart'

Returns the svg markup for the heroicon ''heart'
}
\description{
heart

heart
}
\examples{
rheroicons::outline$heart(
  id = 'my_heart_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the heart icon'
)
rheroicons::solid$heart(
  id = 'my_heart_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the heart icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/heart.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/heart.svg}
}
\keyword{heart}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
