% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{home}
\alias{home}
\title{home}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''home'

Returns the svg markup for the heroicon ''home'
}
\description{
home

home
}
\examples{
rheroicons::outline$home(
  id = 'my_home_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the home icon'
)
rheroicons::solid$home(
  id = 'my_home_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the home icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/home.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/home.svg}
}
\keyword{home}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
