% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\docType{data}
\name{icons}
\alias{icons}
\title{Heroicons for R}
\format{
An object of class \code{list} of length 176.
}
\usage{
icons
}
\description{
Heroicons for R
}
\examples{
rheroicons::icons$book_open()
rheroicons::icons$book_open(type = "outline")
rheroicons::icons$book_open(type = "solid")
rheroicons::icons$book_open(id = "myBookIcon")
rheroicons::icons$book_open(class = "my-icon-set")
rheroicons::icons$book_open(aria_hidden = FALSE)
rheroicons::icons$book_open(title = "read documentation")

}
\references{
\url{https://github.com/refactoringui/heroicons}
\url{https://davidruvolo.shinyapps.io/rheroicons-demo/}
}
\keyword{rheroicons}
