% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{inbox_in}
\alias{inbox_in}
\title{inbox_in}
\usage{
outline$inbox_in()

solid$inbox_in()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a inbox_in

create a solid or outline SVG icon of a inbox_in
}
\description{
inbox_in

inbox_in
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/inbox_in.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/inbox_in.svg}
\url{https://heroicons.dev}
}
\keyword{inbox_in}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
