% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{information_circle}
\alias{information_circle}
\title{information_circle}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''information_circle'

Returns the svg markup for the heroicon ''information_circle'
}
\description{
information_circle

information_circle
}
\examples{
rheroicons::outline$information_circle(
  id = 'my_information_circle_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the information_circle icon'
)
rheroicons::solid$information_circle(
  id = 'my_information_circle_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the information_circle icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/information_circle.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/information_circle.svg}
}
\keyword{information_circle}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
