% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{key}
\alias{key}
\title{key}
\usage{
outline$key()

solid$key()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a key

create a solid or outline SVG icon of a key
}
\description{
key

key
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/key.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/key.svg}
\url{https://heroicons.dev}
}
\keyword{key}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
