% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_gallery.R
\name{launch_gallery}
\alias{launch_gallery}
\title{rheroicons gallery}
\usage{
launch_gallery(...)
}
\arguments{
\item{...}{additional options passed down to \code{shiny::runApp}}
}
\description{
Run the rheroicons gallery and search for icons. Click on an icon
name to copy the function to the clipboard.
}
\examples{
\dontrun{
  rheroicons::launch_gallery()
}

}
