% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{library}
\alias{library}
\title{library}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''library'

Returns the svg markup for the heroicon ''library'
}
\description{
library

library
}
\examples{
rheroicons::outline$library(
  id = 'my_library_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the library icon'
)
rheroicons::solid$library(
  id = 'my_library_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the library icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/library.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/library.svg}
}
\keyword{library}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
