% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{link}
\alias{link}
\title{link}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''link'

Returns the svg markup for the heroicon ''link'
}
\description{
link

link
}
\examples{
rheroicons::outline$link(
  id = 'my_link_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the link icon'
)
rheroicons::solid$link(
  id = 'my_link_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the link icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/link.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/link.svg}
}
\keyword{link}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
