% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{lock_closed}
\alias{lock_closed}
\title{lock_closed}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''lock_closed'

Returns the svg markup for the heroicon ''lock_closed'
}
\description{
lock_closed

lock_closed
}
\examples{
rheroicons::outline$lock_closed(
  id = 'my_lock_closed_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the lock_closed icon'
)
rheroicons::solid$lock_closed(
  id = 'my_lock_closed_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the lock_closed icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/lock_closed.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/lock_closed.svg}
}
\keyword{lock_closed}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
