% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{logout}
\alias{logout}
\title{logout}
\usage{
outline$logout()

solid$logout()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a logout

create a solid or outline SVG icon of a logout
}
\description{
logout

logout
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/logout.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/logout.svg}
\url{https://heroicons.dev}
}
\keyword{logout}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
