% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{mail}
\alias{mail}
\title{mail}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''mail'

Returns the svg markup for the heroicon ''mail'
}
\description{
mail

mail
}
\examples{
rheroicons::outline$mail(
  id = 'my_mail_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the mail icon'
)
rheroicons::solid$mail(
  id = 'my_mail_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the mail icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/mail.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/mail.svg}
}
\keyword{mail}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
