% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{mail_open}
\alias{mail_open}
\title{mail_open}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''mail_open'

Returns the svg markup for the heroicon ''mail_open'
}
\description{
mail_open

mail_open
}
\examples{
rheroicons::outline$mail_open(
  id = 'my_mail_open_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the mail_open icon'
)
rheroicons::solid$mail_open(
  id = 'my_mail_open_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the mail_open icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/mail_open.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/mail_open.svg}
}
\keyword{mail_open}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
