% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{map}
\alias{map}
\title{map}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a map
}
\description{
Render an svg icon of a map
}
\examples{
rheroicons::icons$map(
  type = "solid",
  id = "my_map_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the map icon"
)
}
\keyword{map}
\keyword{rheroicons}
