% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{menu}
\alias{menu}
\title{menu}
\usage{
outline$menu()

solid$menu()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''menu'

Returns the svg markup for the heroicon ''menu'
}
\description{
menu

menu
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/menu.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/menu.svg}
}
\keyword{menu}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
