% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{menu_alt_3}
\alias{menu_alt_3}
\title{menu_alt_3}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''menu_alt_3'

Returns the svg markup for the heroicon ''menu_alt_3'
}
\description{
menu_alt_3

menu_alt_3
}
\examples{
rheroicons::outline$menu_alt_3(
  id = 'my_menu_alt_3_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the menu_alt_3 icon'
)
rheroicons::solid$menu_alt_3(
  id = 'my_menu_alt_3_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the menu_alt_3 icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/menu_alt_3.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/menu_alt_3.svg}
}
\keyword{menu_alt_3}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
