% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{microphone}
\alias{microphone}
\title{microphone}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a microphone
}
\description{
Render an svg icon of a microphone
}
\examples{
rheroicons::icons$microphone(
  type = "solid",
  id = "my_microphone_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the microphone icon"
)
}
\keyword{microphone}
\keyword{rheroicons}
