% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{moon}
\alias{moon}
\title{moon}
\usage{
outline$moon()

solid$moon()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''moon'

Returns the svg markup for the heroicon ''moon'
}
\description{
moon

moon
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/moon.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/moon.svg}
}
\keyword{moon}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
