% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R
\docType{data}
\name{outline}
\alias{outline}
\title{OUTLINE SVG Icons}
\format{
An object of class \code{list} of length 176.
}
\usage{
outline
}
\value{
outline heroicons
}
\description{
OUTLINE SVG Icons
}
\examples{
rheroicons::outline$book_open()
rheroicons::outline$book_open(id = 'myBookIcon')
rheroicons::outline$book_open(class = 'my-icon-set')
rheroicons::outline$book_open(aria_hidden = FALSE, title = 'read document')
}
\references{
\url{https://github.com/refactoringui/heroicons}
\url{https://davidruvolo.shinyapps.io/rheroicons-demo/}
}
\keyword{outline}
\keyword{rheroicons}
