% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{pencil_alt}
\alias{pencil_alt}
\title{pencil_alt}
\usage{
outline$pencil_alt()

solid$pencil_alt()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''pencil_alt'

Returns the svg markup for the heroicon ''pencil_alt'
}
\description{
pencil_alt

pencil_alt
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/pencil_alt.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/pencil_alt.svg}
}
\keyword{outline}
\keyword{pencil_alt}
\keyword{rheroicons}
\keyword{solid}
