% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{phone}
\alias{phone}
\title{phone}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''phone'

Returns the svg markup for the heroicon ''phone'
}
\description{
phone

phone
}
\examples{
rheroicons::outline$phone(
  id = 'my_phone_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the phone icon'
)
rheroicons::solid$phone(
  id = 'my_phone_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the phone icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/phone.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/phone.svg}
}
\keyword{outline}
\keyword{phone}
\keyword{rheroicons}
\keyword{solid}
