% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{phone_incoming}
\alias{phone_incoming}
\title{phone_incoming}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''phone_incoming'

Returns the svg markup for the heroicon ''phone_incoming'
}
\description{
phone_incoming

phone_incoming
}
\examples{
rheroicons::outline$phone_incoming(
  id = 'my_phone_incoming_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the phone_incoming icon'
)
rheroicons::solid$phone_incoming(
  id = 'my_phone_incoming_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the phone_incoming icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/phone_incoming.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/phone_incoming.svg}
}
\keyword{outline}
\keyword{phone_incoming}
\keyword{rheroicons}
\keyword{solid}
