% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{phone_missed_call}
\alias{phone_missed_call}
\title{phone_missed_call}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a phone_missed_call
}
\description{
Render an svg icon of a phone_missed_call
}
\examples{
rheroicons::icons$phone_missed_call(
  type = "solid",
  id = "my_phone_missed_call_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the phone_missed_call icon"
)
}
\keyword{phone_missed_call}
\keyword{rheroicons}
