% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{presentation_chart_bar}
\alias{presentation_chart_bar}
\title{presentation_chart_bar}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a presentation_chart_bar
}
\description{
Render an svg icon of a presentation_chart_bar
}
\examples{
rheroicons::icons$presentation_chart_bar(
  type = "solid",
  id = "my_presentation_chart_bar_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the presentation_chart_bar icon"
)
}
\keyword{presentation_chart_bar}
\keyword{rheroicons}
