% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{puzzle}
\alias{puzzle}
\title{puzzle}
\usage{
outline$puzzle()

solid$puzzle()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''puzzle'

Returns the svg markup for the heroicon ''puzzle'
}
\description{
puzzle

puzzle
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/puzzle.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/puzzle.svg}
}
\keyword{outline}
\keyword{puzzle}
\keyword{rheroicons}
\keyword{solid}
