% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{qrcode}
\alias{qrcode}
\title{qrcode}
\usage{
outline$qrcode()

solid$qrcode()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a qrcode

create a solid or outline SVG icon of a qrcode
}
\description{
qrcode

qrcode
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/qrcode.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/qrcode.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{qrcode}
\keyword{rheroicons}
\keyword{solid}
