% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{question_mark_circle}
\alias{question_mark_circle}
\title{question_mark_circle}
\usage{
outline$question_mark_circle()

solid$question_mark_circle()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a question_mark_circle

create a solid or outline SVG icon of a question_mark_circle
}
\description{
question_mark_circle

question_mark_circle
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/question_mark_circle.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/question_mark_circle.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{question_mark_circle}
\keyword{rheroicons}
\keyword{solid}
