% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{receipt_refund}
\alias{receipt_refund}
\title{receipt_refund}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''receipt_refund'

Returns the svg markup for the heroicon ''receipt_refund'
}
\description{
receipt_refund

receipt_refund
}
\examples{
rheroicons::outline$receipt_refund(
  id = 'my_receipt_refund_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the receipt_refund icon'
)
rheroicons::solid$receipt_refund(
  id = 'my_receipt_refund_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the receipt_refund icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/receipt_refund.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/receipt_refund.svg}
}
\keyword{outline}
\keyword{receipt_refund}
\keyword{rheroicons}
\keyword{solid}
