% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{reply}
\alias{reply}
\title{reply}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''reply'

Returns the svg markup for the heroicon ''reply'
}
\description{
reply

reply
}
\examples{
rheroicons::outline$reply(
  id = 'my_reply_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the reply icon'
)
rheroicons::solid$reply(
  id = 'my_reply_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the reply icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/reply.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/reply.svg}
}
\keyword{outline}
\keyword{reply}
\keyword{rheroicons}
\keyword{solid}
