% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{scale}
\alias{scale}
\title{scale}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''scale'

Returns the svg markup for the heroicon ''scale'
}
\description{
scale

scale
}
\examples{
rheroicons::outline$scale(
  id = 'my_scale_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the scale icon'
)
rheroicons::solid$scale(
  id = 'my_scale_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the scale icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/scale.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/scale.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{scale}
\keyword{solid}
