% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{search}
\alias{search}
\title{search}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''search'

Returns the svg markup for the heroicon ''search'
}
\description{
search

search
}
\examples{
rheroicons::outline$search(
  id = 'my_search_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the search icon'
)
rheroicons::solid$search(
  id = 'my_search_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the search icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/search.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/search.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{search}
\keyword{solid}
