% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{selector}
\alias{selector}
\title{selector}
\usage{
outline$selector()

solid$selector()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a selector

create a solid or outline SVG icon of a selector
}
\description{
selector

selector
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/selector.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/selector.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{selector}
\keyword{solid}
