% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{shield_exclamation}
\alias{shield_exclamation}
\title{shield_exclamation}
\usage{
outline$shield_exclamation()

solid$shield_exclamation()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a shield_exclamation

create a solid or outline SVG icon of a shield_exclamation
}
\description{
shield_exclamation

shield_exclamation
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/shield_exclamation.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/shield_exclamation.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{shield_exclamation}
\keyword{solid}
