% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{shopping_bag}
\alias{shopping_bag}
\title{shopping_bag}
\usage{
outline$shopping_bag()

solid$shopping_bag()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a shopping_bag

create a solid or outline SVG icon of a shopping_bag
}
\description{
shopping_bag

shopping_bag
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/shopping_bag.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/shopping_bag.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{shopping_bag}
\keyword{solid}
