% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solid.R
\docType{data}
\name{solid}
\alias{solid}
\title{SOLID SVG Icons}
\format{
An object of class \code{list} of length 176.
}
\usage{
solid
}
\value{
solid heroicons
}
\description{
SOLID SVG Icons
}
\examples{
rheroicons::solid$book_open()
rheroicons::solid$book_open(id = 'myBookIcon')
rheroicons::solid$book_open(class = 'my-icon-set')
rheroicons::solid$book_open(aria_hidden = FALSE, title = 'read document')
}
\references{
\url{https://github.com/refactoringui/heroicons}
\url{https://davidruvolo.shinyapps.io/rheroicons-demo/}
}
\keyword{rheroicons}
\keyword{solid}
