% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{sort_ascending}
\alias{sort_ascending}
\title{sort_ascending}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''sort_ascending'

Returns the svg markup for the heroicon ''sort_ascending'
}
\description{
sort_ascending

sort_ascending
}
\examples{
rheroicons::outline$sort_ascending(
  id = 'my_sort_ascending_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the sort_ascending icon'
)
rheroicons::solid$sort_ascending(
  id = 'my_sort_ascending_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the sort_ascending icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/sort_ascending.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/sort_ascending.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{sort_ascending}
