% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{sort_descending}
\alias{sort_descending}
\title{sort_descending}
\usage{
outline$sort_descending()

solid$sort_descending()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a sort_descending

create a solid or outline SVG icon of a sort_descending
}
\description{
sort_descending

sort_descending
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/sort_descending.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/sort_descending.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{sort_descending}
