% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{sparkles}
\alias{sparkles}
\title{sparkles}
\usage{
outline$sparkles()

solid$sparkles()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''sparkles'

Returns the svg markup for the heroicon ''sparkles'
}
\description{
sparkles

sparkles
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/sparkles.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/sparkles.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{sparkles}
