% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{star}
\alias{star}
\title{star}
\usage{
outline$star()

solid$star()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''star'

Returns the svg markup for the heroicon ''star'
}
\description{
star

star
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/star.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/star.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{star}
