% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{sun}
\alias{sun}
\title{sun}
\usage{
outline$sun()

solid$sun()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''sun'

Returns the svg markup for the heroicon ''sun'
}
\description{
sun

sun
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/sun.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/sun.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{sun}
