% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{support}
\alias{support}
\title{support}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''support'

Returns the svg markup for the heroicon ''support'
}
\description{
support

support
}
\examples{
rheroicons::outline$support(
  id = 'my_support_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the support icon'
)
rheroicons::solid$support(
  id = 'my_support_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the support icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/support.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/support.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{support}
