% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{tag}
\alias{tag}
\title{tag}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''tag'

Returns the svg markup for the heroicon ''tag'
}
\description{
tag

tag
}
\examples{
rheroicons::outline$tag(
  id = 'my_tag_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the tag icon'
)
rheroicons::solid$tag(
  id = 'my_tag_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the tag icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/tag.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/tag.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{tag}
