% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{terminal}
\alias{terminal}
\title{terminal}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''terminal'

Returns the svg markup for the heroicon ''terminal'
}
\description{
terminal

terminal
}
\examples{
rheroicons::outline$terminal(
  id = 'my_terminal_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the terminal icon'
)
rheroicons::solid$terminal(
  id = 'my_terminal_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the terminal icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/terminal.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/terminal.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{terminal}
