% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{thumb_down}
\alias{thumb_down}
\title{thumb_down}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a thumb_down
}
\description{
Render an svg icon of a thumb_down
}
\examples{
rheroicons::icons$thumb_down(
  type = "solid",
  id = "my_thumb_down_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the thumb_down icon"
)
}
\keyword{rheroicons}
\keyword{thumb_down}
