% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{ticket}
\alias{ticket}
\title{ticket}
\usage{
outline$ticket()

solid$ticket()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''ticket'

Returns the svg markup for the heroicon ''ticket'
}
\description{
ticket

ticket
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/ticket.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/ticket.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{ticket}
