% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{translate}
\alias{translate}
\title{translate}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''translate'

Returns the svg markup for the heroicon ''translate'
}
\description{
translate

translate
}
\examples{
rheroicons::outline$translate(
  id = 'my_translate_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the translate icon'
)
rheroicons::solid$translate(
  id = 'my_translate_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the translate icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/translate.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/translate.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{translate}
