% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{trending_up}
\alias{trending_up}
\title{trending_up}
\usage{
outline$trending_up()

solid$trending_up()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''trending_up'

Returns the svg markup for the heroicon ''trending_up'
}
\description{
trending_up

trending_up
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/trending_up.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/trending_up.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{trending_up}
