% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{upload}
\alias{upload}
\title{upload}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''upload'

Returns the svg markup for the heroicon ''upload'
}
\description{
upload

upload
}
\examples{
rheroicons::outline$upload(
  id = 'my_upload_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the upload icon'
)
rheroicons::solid$upload(
  id = 'my_upload_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the upload icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/upload.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/upload.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{upload}
