% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{user}
\alias{user}
\title{user}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''user'

Returns the svg markup for the heroicon ''user'
}
\description{
user

user
}
\examples{
rheroicons::outline$user(
  id = 'my_user_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the user icon'
)
rheroicons::solid$user(
  id = 'my_user_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the user icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/user.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/user.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{user}
