% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{user_add}
\alias{user_add}
\title{user_add}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a user_add
}
\description{
Render an svg icon of a user_add
}
\examples{
rheroicons::icons$user_add(
  type = "solid",
  id = "my_user_add_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the user_add icon"
)
}
\keyword{rheroicons}
\keyword{user_add}
