% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{user_circle}
\alias{user_circle}
\title{user_circle}
\usage{
outline$user_circle()

solid$user_circle()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''user_circle'

Returns the svg markup for the heroicon ''user_circle'
}
\description{
user_circle

user_circle
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/user_circle.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/user_circle.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{user_circle}
