% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{users}
\alias{users}
\title{users}
\usage{
outline$users()

solid$users()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''users'

Returns the svg markup for the heroicon ''users'
}
\description{
users

users
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/users.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/users.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{users}
