% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{view_boards}
\alias{view_boards}
\title{view_boards}
\usage{
outline$view_boards()

solid$view_boards()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a view_boards

create a solid or outline SVG icon of a view_boards
}
\description{
view_boards

view_boards
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/view_boards.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/view_boards.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{view_boards}
