% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{view_grid}
\alias{view_grid}
\title{view_grid}
\usage{
outline$view_grid()

solid$view_grid()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a view_grid

create a solid or outline SVG icon of a view_grid
}
\description{
view_grid

view_grid
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/view_grid.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/view_grid.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{view_grid}
