% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{view_grid_add}
\alias{view_grid_add}
\title{view_grid_add}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''view_grid_add'

Returns the svg markup for the heroicon ''view_grid_add'
}
\description{
view_grid_add

view_grid_add
}
\examples{
rheroicons::outline$view_grid_add(
  id = 'my_view_grid_add_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the view_grid_add icon'
)
rheroicons::solid$view_grid_add(
  id = 'my_view_grid_add_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the view_grid_add icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/view_grid_add.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/view_grid_add.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{view_grid_add}
