% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{volume_off}
\alias{volume_off}
\title{volume_off}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''volume_off'

Returns the svg markup for the heroicon ''volume_off'
}
\description{
volume_off

volume_off
}
\examples{
rheroicons::outline$volume_off(
  id = 'my_volume_off_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the volume_off icon'
)
rheroicons::solid$volume_off(
  id = 'my_volume_off_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the volume_off icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/volume_off.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/volume_off.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{volume_off}
