% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{x_circle}
\alias{x_circle}
\title{x_circle}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''x_circle'

Returns the svg markup for the heroicon ''x_circle'
}
\description{
x_circle

x_circle
}
\examples{
rheroicons::outline$x_circle(
  id = 'my_x_circle_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the x_circle icon'
)
rheroicons::solid$x_circle(
  id = 'my_x_circle_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the x_circle icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/x_circle.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/x_circle.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{x_circle}
