% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{zoom_out}
\alias{zoom_out}
\title{zoom_out}
\usage{
outline$zoom_out()

solid$zoom_out()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''zoom_out'

Returns the svg markup for the heroicon ''zoom_out'
}
\description{
zoom_out

zoom_out
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/zoom_out.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/zoom_out.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{zoom_out}
