% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{adjustments}
\alias{adjustments}
\title{adjustments}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''adjustments'

Returns the svg markup for the heroicon ''adjustments'
}
\description{
adjustments

adjustments
}
\examples{
rheroicons::outline$adjustments(
  id = 'my_adjustments_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the adjustments icon'
)
rheroicons::solid$adjustments(
  id = 'my_adjustments_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the adjustments icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/adjustments.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/adjustments.svg}
}
\keyword{adjustments}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
