% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{annotation}
\alias{annotation}
\title{annotation}
\usage{
outline$annotation()

solid$annotation()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''annotation'

Returns the svg markup for the heroicon ''annotation'
}
\description{
annotation

annotation
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/annotation.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/annotation.svg}
}
\keyword{annotation}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
