% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{archive}
\alias{archive}
\title{archive}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''archive'

Returns the svg markup for the heroicon ''archive'
}
\description{
archive

archive
}
\examples{
rheroicons::outline$archive(
  id = 'my_archive_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the archive icon'
)
rheroicons::solid$archive(
  id = 'my_archive_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the archive icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/archive.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/archive.svg}
}
\keyword{archive}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
