% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{archive}
\alias{archive}
\title{archive}
\usage{
outline$archive()

solid$archive()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a archive

create a solid or outline SVG icon of a archive
}
\description{
archive

archive
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/archive.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/archive.svg}
\url{https://heroicons.dev}
}
\keyword{archive}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
