% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{arrow_circle_down}
\alias{arrow_circle_down}
\title{arrow_circle_down}
\usage{
outline$arrow_circle_down()

solid$arrow_circle_down()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''arrow_circle_down'

Returns the svg markup for the heroicon ''arrow_circle_down'
}
\description{
arrow_circle_down

arrow_circle_down
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/arrow_circle_down.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/arrow_circle_down.svg}
}
\keyword{arrow_circle_down}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
