% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{arrow_circle_left}
\alias{arrow_circle_left}
\title{arrow_circle_left}
\usage{
outline$arrow_circle_left()

solid$arrow_circle_left()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a arrow_circle_left

create a solid or outline SVG icon of a arrow_circle_left
}
\description{
arrow_circle_left

arrow_circle_left
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/arrow_circle_left.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/arrow_circle_left.svg}
\url{https://heroicons.dev}
}
\keyword{arrow_circle_left}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
