% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{arrow_circle_right}
\alias{arrow_circle_right}
\title{arrow_circle_right}
\usage{
outline$arrow_circle_right()

solid$arrow_circle_right()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''arrow_circle_right'

Returns the svg markup for the heroicon ''arrow_circle_right'
}
\description{
arrow_circle_right

arrow_circle_right
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/arrow_circle_right.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/arrow_circle_right.svg}
}
\keyword{arrow_circle_right}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
