% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{arrow_circle_up}
\alias{arrow_circle_up}
\title{arrow_circle_up}
\usage{
outline$arrow_circle_up()

solid$arrow_circle_up()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a arrow_circle_up

create a solid or outline SVG icon of a arrow_circle_up
}
\description{
arrow_circle_up

arrow_circle_up
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/arrow_circle_up.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/arrow_circle_up.svg}
\url{https://heroicons.dev}
}
\keyword{arrow_circle_up}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
