% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{arrow_left}
\alias{arrow_left}
\title{arrow_left}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''arrow_left'

Returns the svg markup for the heroicon ''arrow_left'
}
\description{
arrow_left

arrow_left
}
\examples{
rheroicons::outline$arrow_left(
  id = 'my_arrow_left_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the arrow_left icon'
)
rheroicons::solid$arrow_left(
  id = 'my_arrow_left_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the arrow_left icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/arrow_left.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/arrow_left.svg}
}
\keyword{arrow_left}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
