% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{arrow_left}
\alias{arrow_left}
\title{arrow_left}
\usage{
outline$arrow_left()

solid$arrow_left()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''arrow_left'

Returns the svg markup for the heroicon ''arrow_left'
}
\description{
arrow_left

arrow_left
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/arrow_left.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/arrow_left.svg}
}
\keyword{arrow_left}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
