% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{arrow_narrow_right}
\alias{arrow_narrow_right}
\title{arrow_narrow_right}
\usage{
outline$arrow_narrow_right()

solid$arrow_narrow_right()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a arrow_narrow_right

create a solid or outline SVG icon of a arrow_narrow_right
}
\description{
arrow_narrow_right

arrow_narrow_right
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/arrow_narrow_right.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/arrow_narrow_right.svg}
\url{https://heroicons.dev}
}
\keyword{arrow_narrow_right}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
