% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{arrow_up}
\alias{arrow_up}
\title{arrow_up}
\usage{
outline$arrow_up()

solid$arrow_up()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''arrow_up'

Returns the svg markup for the heroicon ''arrow_up'
}
\description{
arrow_up

arrow_up
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/arrow_up.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/arrow_up.svg}
}
\keyword{arrow_up}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
