% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{arrows_expand}
\alias{arrows_expand}
\title{arrows_expand}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''arrows_expand'

Returns the svg markup for the heroicon ''arrows_expand'
}
\description{
arrows_expand

arrows_expand
}
\examples{
rheroicons::outline$arrows_expand(
  id = 'my_arrows_expand_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the arrows_expand icon'
)
rheroicons::solid$arrows_expand(
  id = 'my_arrows_expand_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the arrows_expand icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/arrows_expand.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/arrows_expand.svg}
}
\keyword{arrows_expand}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
