% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{arrows_expand}
\alias{arrows_expand}
\title{arrows_expand}
\usage{
outline$arrows_expand()

solid$arrows_expand()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''arrows_expand'

Returns the svg markup for the heroicon ''arrows_expand'
}
\description{
arrows_expand

arrows_expand
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/arrows_expand.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/arrows_expand.svg}
}
\keyword{arrows_expand}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
