% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{at_symbol}
\alias{at_symbol}
\title{at_symbol}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''at_symbol'

Returns the svg markup for the heroicon ''at_symbol'
}
\description{
at_symbol

at_symbol
}
\examples{
rheroicons::outline$at_symbol(
  id = 'my_at_symbol_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the at_symbol icon'
)
rheroicons::solid$at_symbol(
  id = 'my_at_symbol_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the at_symbol icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/at_symbol.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/at_symbol.svg}
}
\keyword{at_symbol}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
