% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{badge_check}
\alias{badge_check}
\title{badge_check}
\usage{
outline$badge_check()

solid$badge_check()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a badge_check

create a solid or outline SVG icon of a badge_check
}
\description{
badge_check

badge_check
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/badge_check.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/badge_check.svg}
\url{https://heroicons.dev}
}
\keyword{badge_check}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
