% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{badge_check}
\alias{badge_check}
\title{badge_check}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''badge_check'

Returns the svg markup for the heroicon ''badge_check'
}
\description{
badge_check

badge_check
}
\examples{
rheroicons::outline$badge_check(
  id = 'my_badge_check_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the badge_check icon'
)
rheroicons::solid$badge_check(
  id = 'my_badge_check_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the badge_check icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/badge_check.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/badge_check.svg}
}
\keyword{badge_check}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
