% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{bell}
\alias{bell}
\title{bell}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''bell'

Returns the svg markup for the heroicon ''bell'
}
\description{
bell

bell
}
\examples{
rheroicons::outline$bell(
  id = 'my_bell_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the bell icon'
)
rheroicons::solid$bell(
  id = 'my_bell_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the bell icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/bell.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/bell.svg}
}
\keyword{bell}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
