% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{book_open}
\alias{book_open}
\title{book_open}
\usage{
outline$book_open()

solid$book_open()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a book_open

create a solid or outline SVG icon of a book_open
}
\description{
book_open

book_open
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/book_open.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/book_open.svg}
\url{https://heroicons.dev}
}
\keyword{book_open}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
