% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{book_open}
\alias{book_open}
\title{book_open}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a book_open
}
\description{
Render an svg icon of a book_open
}
\examples{
rheroicons::icons$book_open(
  type = "solid",
  id = "my_book_open_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the book_open icon"
)
}
\keyword{book_open}
\keyword{rheroicons}
