% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{bookmark}
\alias{bookmark}
\title{bookmark}
\usage{
outline$bookmark()

solid$bookmark()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''bookmark'

Returns the svg markup for the heroicon ''bookmark'
}
\description{
bookmark

bookmark
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/bookmark.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/bookmark.svg}
}
\keyword{bookmark}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
