% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{bookmark_alt}
\alias{bookmark_alt}
\title{bookmark_alt}
\usage{
outline$bookmark_alt()

solid$bookmark_alt()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''bookmark_alt'

Returns the svg markup for the heroicon ''bookmark_alt'
}
\description{
bookmark_alt

bookmark_alt
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/bookmark_alt.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/bookmark_alt.svg}
}
\keyword{bookmark_alt}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
