% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{briefcase}
\alias{briefcase}
\title{briefcase}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''briefcase'

Returns the svg markup for the heroicon ''briefcase'
}
\description{
briefcase

briefcase
}
\examples{
rheroicons::outline$briefcase(
  id = 'my_briefcase_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the briefcase icon'
)
rheroicons::solid$briefcase(
  id = 'my_briefcase_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the briefcase icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/briefcase.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/briefcase.svg}
}
\keyword{briefcase}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
