% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{calendar}
\alias{calendar}
\title{calendar}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''calendar'

Returns the svg markup for the heroicon ''calendar'
}
\description{
calendar

calendar
}
\examples{
rheroicons::outline$calendar(
  id = 'my_calendar_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the calendar icon'
)
rheroicons::solid$calendar(
  id = 'my_calendar_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the calendar icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/calendar.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/calendar.svg}
}
\keyword{calendar}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
