% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{calendar}
\alias{calendar}
\title{calendar}
\usage{
outline$calendar()

solid$calendar()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''calendar'

Returns the svg markup for the heroicon ''calendar'
}
\description{
calendar

calendar
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/calendar.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/calendar.svg}
}
\keyword{calendar}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
