% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{camera}
\alias{camera}
\title{camera}
\usage{
outline$camera()

solid$camera()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''camera'

Returns the svg markup for the heroicon ''camera'
}
\description{
camera

camera
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/camera.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/camera.svg}
}
\keyword{camera}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
