% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{cash}
\alias{cash}
\title{cash}
\usage{
outline$cash()

solid$cash()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a cash

create a solid or outline SVG icon of a cash
}
\description{
cash

cash
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/cash.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/cash.svg}
\url{https://heroicons.dev}
}
\keyword{cash}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
