% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{cash}
\alias{cash}
\title{cash}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''cash'

Returns the svg markup for the heroicon ''cash'
}
\description{
cash

cash
}
\examples{
rheroicons::outline$cash(
  id = 'my_cash_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the cash icon'
)
rheroicons::solid$cash(
  id = 'my_cash_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the cash icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/cash.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/cash.svg}
}
\keyword{cash}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
