% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{chart_square_bar}
\alias{chart_square_bar}
\title{chart_square_bar}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a chart_square_bar
}
\description{
Render an svg icon of a chart_square_bar
}
\examples{
rheroicons::icons$chart_square_bar(
  type = "solid",
  id = "my_chart_square_bar_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the chart_square_bar icon"
)
}
\keyword{chart_square_bar}
\keyword{rheroicons}
