% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{chat}
\alias{chat}
\title{chat}
\usage{
outline$chat()

solid$chat()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a chat

create a solid or outline SVG icon of a chat
}
\description{
chat

chat
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/chat.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/chat.svg}
\url{https://heroicons.dev}
}
\keyword{chat}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
