% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{check}
\alias{check}
\title{check}
\usage{
outline$check()

solid$check()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a check

create a solid or outline SVG icon of a check
}
\description{
check

check
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/check.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/check.svg}
\url{https://heroicons.dev}
}
\keyword{check}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
