% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{check}
\alias{check}
\title{check}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''check'

Returns the svg markup for the heroicon ''check'
}
\description{
check

check
}
\examples{
rheroicons::outline$check(
  id = 'my_check_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the check icon'
)
rheroicons::solid$check(
  id = 'my_check_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the check icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/check.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/check.svg}
}
\keyword{check}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
