% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{check_circle}
\alias{check_circle}
\title{check_circle}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''check_circle'

Returns the svg markup for the heroicon ''check_circle'
}
\description{
check_circle

check_circle
}
\examples{
rheroicons::outline$check_circle(
  id = 'my_check_circle_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the check_circle icon'
)
rheroicons::solid$check_circle(
  id = 'my_check_circle_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the check_circle icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/check_circle.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/check_circle.svg}
}
\keyword{check_circle}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
