% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{chevron_down}
\alias{chevron_down}
\title{chevron_down}
\usage{
outline$chevron_down()

solid$chevron_down()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a chevron_down

create a solid or outline SVG icon of a chevron_down
}
\description{
chevron_down

chevron_down
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/chevron_down.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/chevron_down.svg}
\url{https://heroicons.dev}
}
\keyword{chevron_down}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
