% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{chevron_left}
\alias{chevron_left}
\title{chevron_left}
\usage{
outline$chevron_left()

solid$chevron_left()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''chevron_left'

Returns the svg markup for the heroicon ''chevron_left'
}
\description{
chevron_left

chevron_left
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/chevron_left.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/chevron_left.svg}
}
\keyword{chevron_left}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
