% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{clipboard}
\alias{clipboard}
\title{clipboard}
\usage{
outline$clipboard()

solid$clipboard()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''clipboard'

Returns the svg markup for the heroicon ''clipboard'
}
\description{
clipboard

clipboard
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/clipboard.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/clipboard.svg}
}
\keyword{clipboard}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
