% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{clipboard_copy}
\alias{clipboard_copy}
\title{clipboard_copy}
\usage{
outline$clipboard_copy()

solid$clipboard_copy()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a clipboard_copy

create a solid or outline SVG icon of a clipboard_copy
}
\description{
clipboard_copy

clipboard_copy
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/clipboard_copy.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/clipboard_copy.svg}
\url{https://heroicons.dev}
}
\keyword{clipboard_copy}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
