% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{cloud_upload}
\alias{cloud_upload}
\title{cloud_upload}
\usage{
outline$cloud_upload()

solid$cloud_upload()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a cloud_upload

create a solid or outline SVG icon of a cloud_upload
}
\description{
cloud_upload

cloud_upload
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/cloud_upload.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/cloud_upload.svg}
\url{https://heroicons.dev}
}
\keyword{cloud_upload}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
