% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{code}
\alias{code}
\title{code}
\usage{
outline$code()

solid$code()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a code

create a solid or outline SVG icon of a code
}
\description{
code

code
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/code.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/code.svg}
\url{https://heroicons.dev}
}
\keyword{code}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
