% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{code}
\alias{code}
\title{code}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''code'

Returns the svg markup for the heroicon ''code'
}
\description{
code

code
}
\examples{
rheroicons::outline$code(
  id = 'my_code_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the code icon'
)
rheroicons::solid$code(
  id = 'my_code_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the code icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/code.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/code.svg}
}
\keyword{code}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
