% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{cog}
\alias{cog}
\title{cog}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''cog'

Returns the svg markup for the heroicon ''cog'
}
\description{
cog

cog
}
\examples{
rheroicons::outline$cog(
  id = 'my_cog_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the cog icon'
)
rheroicons::solid$cog(
  id = 'my_cog_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the cog icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/cog.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/cog.svg}
}
\keyword{cog}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
