% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{collection}
\alias{collection}
\title{collection}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''collection'

Returns the svg markup for the heroicon ''collection'
}
\description{
collection

collection
}
\examples{
rheroicons::outline$collection(
  id = 'my_collection_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the collection icon'
)
rheroicons::solid$collection(
  id = 'my_collection_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the collection icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/collection.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/collection.svg}
}
\keyword{collection}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
