% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{color_swatch}
\alias{color_swatch}
\title{color_swatch}
\usage{
outline$color_swatch()

solid$color_swatch()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''color_swatch'

Returns the svg markup for the heroicon ''color_swatch'
}
\description{
color_swatch

color_swatch
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/color_swatch.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/color_swatch.svg}
}
\keyword{color_swatch}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
