% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{credit_card}
\alias{credit_card}
\title{credit_card}
\usage{
outline$credit_card()

solid$credit_card()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''credit_card'

Returns the svg markup for the heroicon ''credit_card'
}
\description{
credit_card

credit_card
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/credit_card.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/credit_card.svg}
}
\keyword{credit_card}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
