% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{currency_dollar}
\alias{currency_dollar}
\title{currency_dollar}
\usage{
outline$currency_dollar()

solid$currency_dollar()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a currency_dollar

create a solid or outline SVG icon of a currency_dollar
}
\description{
currency_dollar

currency_dollar
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/currency_dollar.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/currency_dollar.svg}
\url{https://heroicons.dev}
}
\keyword{currency_dollar}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
