% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{currency_euro}
\alias{currency_euro}
\title{currency_euro}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a currency_euro
}
\description{
Render an svg icon of a currency_euro
}
\examples{
rheroicons::icons$currency_euro(
  type = "solid",
  id = "my_currency_euro_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the currency_euro icon"
)
}
\keyword{currency_euro}
\keyword{rheroicons}
