% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{currency_euro}
\alias{currency_euro}
\title{currency_euro}
\usage{
outline$currency_euro()

solid$currency_euro()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a currency_euro

create a solid or outline SVG icon of a currency_euro
}
\description{
currency_euro

currency_euro
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/currency_euro.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/currency_euro.svg}
\url{https://heroicons.dev}
}
\keyword{currency_euro}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
