% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{currency_pound}
\alias{currency_pound}
\title{currency_pound}
\usage{
outline$currency_pound()

solid$currency_pound()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a currency_pound

create a solid or outline SVG icon of a currency_pound
}
\description{
currency_pound

currency_pound
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/currency_pound.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/currency_pound.svg}
\url{https://heroicons.dev}
}
\keyword{currency_pound}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
