% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{currency_rupee}
\alias{currency_rupee}
\title{currency_rupee}
\usage{
outline$currency_rupee()

solid$currency_rupee()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''currency_rupee'

Returns the svg markup for the heroicon ''currency_rupee'
}
\description{
currency_rupee

currency_rupee
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/currency_rupee.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/currency_rupee.svg}
}
\keyword{currency_rupee}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
