% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{currency_yen}
\alias{currency_yen}
\title{currency_yen}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a currency_yen
}
\description{
Render an svg icon of a currency_yen
}
\examples{
rheroicons::icons$currency_yen(
  type = "solid",
  id = "my_currency_yen_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the currency_yen icon"
)
}
\keyword{currency_yen}
\keyword{rheroicons}
