% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{currency_yen}
\alias{currency_yen}
\title{currency_yen}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''currency_yen'

Returns the svg markup for the heroicon ''currency_yen'
}
\description{
currency_yen

currency_yen
}
\examples{
rheroicons::outline$currency_yen(
  id = 'my_currency_yen_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the currency_yen icon'
)
rheroicons::solid$currency_yen(
  id = 'my_currency_yen_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the currency_yen icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/currency_yen.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/currency_yen.svg}
}
\keyword{currency_yen}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
