% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{cursor_click}
\alias{cursor_click}
\title{cursor_click}
\usage{
outline$cursor_click()

solid$cursor_click()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''cursor_click'

Returns the svg markup for the heroicon ''cursor_click'
}
\description{
cursor_click

cursor_click
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/cursor_click.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/cursor_click.svg}
}
\keyword{cursor_click}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
