% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{desktop_computer}
\alias{desktop_computer}
\title{desktop_computer}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a desktop_computer
}
\description{
Render an svg icon of a desktop_computer
}
\examples{
rheroicons::icons$desktop_computer(
  type = "solid",
  id = "my_desktop_computer_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the desktop_computer icon"
)
}
\keyword{desktop_computer}
\keyword{rheroicons}
