% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{document}
\alias{document}
\title{document}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''document'

Returns the svg markup for the heroicon ''document'
}
\description{
document

document
}
\examples{
rheroicons::outline$document(
  id = 'my_document_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the document icon'
)
rheroicons::solid$document(
  id = 'my_document_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the document icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/document.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/document.svg}
}
\keyword{document}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
