% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{document}
\alias{document}
\title{document}
\usage{
outline$document()

solid$document()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''document'

Returns the svg markup for the heroicon ''document'
}
\description{
document

document
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/document.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/document.svg}
}
\keyword{document}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
