% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{document_add}
\alias{document_add}
\title{document_add}
\usage{
outline$document_add()

solid$document_add()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a document_add

create a solid or outline SVG icon of a document_add
}
\description{
document_add

document_add
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/document_add.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/document_add.svg}
\url{https://heroicons.dev}
}
\keyword{document_add}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
