% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{document_download}
\alias{document_download}
\title{document_download}
\usage{
outline$document_download()

solid$document_download()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''document_download'

Returns the svg markup for the heroicon ''document_download'
}
\description{
document_download

document_download
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/document_download.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/document_download.svg}
}
\keyword{document_download}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
