% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{document_duplicate}
\alias{document_duplicate}
\title{document_duplicate}
\usage{
outline$document_duplicate()

solid$document_duplicate()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''document_duplicate'

Returns the svg markup for the heroicon ''document_duplicate'
}
\description{
document_duplicate

document_duplicate
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/document_duplicate.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/document_duplicate.svg}
}
\keyword{document_duplicate}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
