% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{document_remove}
\alias{document_remove}
\title{document_remove}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a document_remove
}
\description{
Render an svg icon of a document_remove
}
\examples{
rheroicons::icons$document_remove(
  type = "solid",
  id = "my_document_remove_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the document_remove icon"
)
}
\keyword{document_remove}
\keyword{rheroicons}
