% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{document_report}
\alias{document_report}
\title{document_report}
\usage{
outline$document_report()

solid$document_report()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a document_report

create a solid or outline SVG icon of a document_report
}
\description{
document_report

document_report
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/document_report.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/document_report.svg}
\url{https://heroicons.dev}
}
\keyword{document_report}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
