% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{dots_circle_horizontal}
\alias{dots_circle_horizontal}
\title{dots_circle_horizontal}
\usage{
outline$dots_circle_horizontal()

solid$dots_circle_horizontal()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''dots_circle_horizontal'

Returns the svg markup for the heroicon ''dots_circle_horizontal'
}
\description{
dots_circle_horizontal

dots_circle_horizontal
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/dots_circle_horizontal.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/dots_circle_horizontal.svg}
}
\keyword{dots_circle_horizontal}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
