% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{dots_horizontal}
\alias{dots_horizontal}
\title{dots_horizontal}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''dots_horizontal'

Returns the svg markup for the heroicon ''dots_horizontal'
}
\description{
dots_horizontal

dots_horizontal
}
\examples{
rheroicons::outline$dots_horizontal(
  id = 'my_dots_horizontal_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the dots_horizontal icon'
)
rheroicons::solid$dots_horizontal(
  id = 'my_dots_horizontal_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the dots_horizontal icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/dots_horizontal.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/dots_horizontal.svg}
}
\keyword{dots_horizontal}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
