% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{dots_vertical}
\alias{dots_vertical}
\title{dots_vertical}
\usage{
outline$dots_vertical()

solid$dots_vertical()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a dots_vertical

create a solid or outline SVG icon of a dots_vertical
}
\description{
dots_vertical

dots_vertical
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/dots_vertical.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/dots_vertical.svg}
\url{https://heroicons.dev}
}
\keyword{dots_vertical}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
