% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{dots_vertical}
\alias{dots_vertical}
\title{dots_vertical}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''dots_vertical'

Returns the svg markup for the heroicon ''dots_vertical'
}
\description{
dots_vertical

dots_vertical
}
\examples{
rheroicons::outline$dots_vertical(
  id = 'my_dots_vertical_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the dots_vertical icon'
)
rheroicons::solid$dots_vertical(
  id = 'my_dots_vertical_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the dots_vertical icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/dots_vertical.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/dots_vertical.svg}
}
\keyword{dots_vertical}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
