% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{download}
\alias{download}
\title{download}
\usage{
outline$download()

solid$download()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''download'

Returns the svg markup for the heroicon ''download'
}
\description{
download

download
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/download.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/download.svg}
}
\keyword{download}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
