% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{duplicate}
\alias{duplicate}
\title{duplicate}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''duplicate'

Returns the svg markup for the heroicon ''duplicate'
}
\description{
duplicate

duplicate
}
\examples{
rheroicons::outline$duplicate(
  id = 'my_duplicate_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the duplicate icon'
)
rheroicons::solid$duplicate(
  id = 'my_duplicate_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the duplicate icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/duplicate.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/duplicate.svg}
}
\keyword{duplicate}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
