% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{emoji_happy}
\alias{emoji_happy}
\title{emoji_happy}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''emoji_happy'

Returns the svg markup for the heroicon ''emoji_happy'
}
\description{
emoji_happy

emoji_happy
}
\examples{
rheroicons::outline$emoji_happy(
  id = 'my_emoji_happy_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the emoji_happy icon'
)
rheroicons::solid$emoji_happy(
  id = 'my_emoji_happy_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the emoji_happy icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/emoji_happy.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/emoji_happy.svg}
}
\keyword{emoji_happy}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
