% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{emoji_sad}
\alias{emoji_sad}
\title{emoji_sad}
\usage{
outline$emoji_sad()

solid$emoji_sad()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a emoji_sad

create a solid or outline SVG icon of a emoji_sad
}
\description{
emoji_sad

emoji_sad
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/emoji_sad.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/emoji_sad.svg}
\url{https://heroicons.dev}
}
\keyword{emoji_sad}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
