% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{exclamation_circle}
\alias{exclamation_circle}
\title{exclamation_circle}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''exclamation_circle'

Returns the svg markup for the heroicon ''exclamation_circle'
}
\description{
exclamation_circle

exclamation_circle
}
\examples{
rheroicons::outline$exclamation_circle(
  id = 'my_exclamation_circle_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the exclamation_circle icon'
)
rheroicons::solid$exclamation_circle(
  id = 'my_exclamation_circle_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the exclamation_circle icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/exclamation_circle.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/exclamation_circle.svg}
}
\keyword{exclamation_circle}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
