% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{exclamation_circle}
\alias{exclamation_circle}
\title{exclamation_circle}
\usage{
outline$exclamation_circle()

solid$exclamation_circle()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a exclamation_circle

create a solid or outline SVG icon of a exclamation_circle
}
\description{
exclamation_circle

exclamation_circle
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/exclamation_circle.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/exclamation_circle.svg}
\url{https://heroicons.dev}
}
\keyword{exclamation_circle}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
