% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{external_link}
\alias{external_link}
\title{external_link}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''external_link'

Returns the svg markup for the heroicon ''external_link'
}
\description{
external_link

external_link
}
\examples{
rheroicons::outline$external_link(
  id = 'my_external_link_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the external_link icon'
)
rheroicons::solid$external_link(
  id = 'my_external_link_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the external_link icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/external_link.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/external_link.svg}
}
\keyword{external_link}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
