% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{external_link}
\alias{external_link}
\title{external_link}
\usage{
outline$external_link()

solid$external_link()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a external_link

create a solid or outline SVG icon of a external_link
}
\description{
external_link

external_link
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/external_link.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/external_link.svg}
\url{https://heroicons.dev}
}
\keyword{external_link}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
