% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{external_link}
\alias{external_link}
\title{external_link}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a external_link
}
\description{
Render an svg icon of a external_link
}
\examples{
rheroicons::icons$external_link(
  type = "solid",
  id = "my_external_link_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the external_link icon"
)
}
\keyword{external_link}
\keyword{rheroicons}
