% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{eye}
\alias{eye}
\title{eye}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a eye
}
\description{
Render an svg icon of a eye
}
\examples{
rheroicons::icons$eye(
  type = "solid",
  id = "my_eye_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the eye icon"
)
}
\keyword{eye}
\keyword{rheroicons}
