% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{eye}
\alias{eye}
\title{eye}
\usage{
outline$eye()

solid$eye()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a eye

create a solid or outline SVG icon of a eye
}
\description{
eye

eye
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/eye.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/eye.svg}
\url{https://heroicons.dev}
}
\keyword{eye}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
