% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{eye_off}
\alias{eye_off}
\title{eye_off}
\usage{
outline$eye_off()

solid$eye_off()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''eye_off'

Returns the svg markup for the heroicon ''eye_off'
}
\description{
eye_off

eye_off
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/eye_off.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/eye_off.svg}
}
\keyword{eye_off}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
