% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_icon.R
\name{find_icons}
\alias{find_icons}
\title{Find available icons}
\usage{
find_icons(query = "")
}
\arguments{
\item{query}{a search term passed down to \code{stringr::str_subset}.
Alternatively, leave blank to view all icon names.}
}
\value{
An array of icon names
}
\description{
Find available icons
}
\examples{
find_icons(query = "chevron")
find_icons(query = "chevron|arrow")
find_icons(query = "down|up|left|right")

}
\references{
\url{https://heroicons.com}
\url{https://github.com/tailwindlabs/heroicons}
}
