% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_icon.R
\name{find_icons}
\alias{find_icons}
\title{Find available icons}
\usage{
find_icons(query = "")
}
\arguments{
\item{query}{a string containing a search term}
}
\value{
An array of icon names
}
\description{
You can search for icons by querying the icon set. For example, if you would
like to find icons that are arrows and chevrons, then type in you query as a
regexp: 'arrows|chevrons'. Alternatively, if you would like to view all icon
names, leave the value for 'query' blank.
}
\examples{
find_icons(query = "chevron")
find_icons(query = "chevron|arrow")
find_icons(query = "down|up|left|right")
find_icons(query = "_circle_")
find_icons(query = "(\\\\_down)$")

}
\references{
\url{https://heroicons.com}
\url{https://github.com/tailwindlabs/heroicons}
}
