% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{flag}
\alias{flag}
\title{flag}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a flag
}
\description{
Render an svg icon of a flag
}
\examples{
rheroicons::icons$flag(
  type = "solid",
  id = "my_flag_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the flag icon"
)
}
\keyword{flag}
\keyword{rheroicons}
