% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{flag}
\alias{flag}
\title{flag}
\usage{
outline$flag()

solid$flag()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''flag'

Returns the svg markup for the heroicon ''flag'
}
\description{
flag

flag
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/flag.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/flag.svg}
}
\keyword{flag}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
