% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{folder}
\alias{folder}
\title{folder}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''folder'

Returns the svg markup for the heroicon ''folder'
}
\description{
folder

folder
}
\examples{
rheroicons::outline$folder(
  id = 'my_folder_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the folder icon'
)
rheroicons::solid$folder(
  id = 'my_folder_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the folder icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/folder.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/folder.svg}
}
\keyword{folder}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
