% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{folder}
\alias{folder}
\title{folder}
\usage{
outline$folder()

solid$folder()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a folder

create a solid or outline SVG icon of a folder
}
\description{
folder

folder
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/folder.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/folder.svg}
\url{https://heroicons.dev}
}
\keyword{folder}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
