% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{folder_add}
\alias{folder_add}
\title{folder_add}
\usage{
outline$folder_add()

solid$folder_add()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''folder_add'

Returns the svg markup for the heroicon ''folder_add'
}
\description{
folder_add

folder_add
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/folder_add.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/folder_add.svg}
}
\keyword{folder_add}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
