% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{folder_download}
\alias{folder_download}
\title{folder_download}
\usage{
outline$folder_download()

solid$folder_download()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''folder_download'

Returns the svg markup for the heroicon ''folder_download'
}
\description{
folder_download

folder_download
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/folder_download.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/folder_download.svg}
}
\keyword{folder_download}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
