% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{folder_remove}
\alias{folder_remove}
\title{folder_remove}
\usage{
outline$folder_remove()

solid$folder_remove()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''folder_remove'

Returns the svg markup for the heroicon ''folder_remove'
}
\description{
folder_remove

folder_remove
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/folder_remove.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/folder_remove.svg}
}
\keyword{folder_remove}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
