% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{globe}
\alias{globe}
\title{globe}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''globe'

Returns the svg markup for the heroicon ''globe'
}
\description{
globe

globe
}
\examples{
rheroicons::outline$globe(
  id = 'my_globe_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the globe icon'
)
rheroicons::solid$globe(
  id = 'my_globe_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the globe icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/globe.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/globe.svg}
}
\keyword{globe}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
