% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{globe_alt}
\alias{globe_alt}
\title{globe_alt}
\usage{
outline$globe_alt()

solid$globe_alt()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a globe_alt

create a solid or outline SVG icon of a globe_alt
}
\description{
globe_alt

globe_alt
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/globe_alt.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/globe_alt.svg}
\url{https://heroicons.dev}
}
\keyword{globe_alt}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
