% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{globe_alt}
\alias{globe_alt}
\title{globe_alt}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''globe_alt'

Returns the svg markup for the heroicon ''globe_alt'
}
\description{
globe_alt

globe_alt
}
\examples{
rheroicons::outline$globe_alt(
  id = 'my_globe_alt_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the globe_alt icon'
)
rheroicons::solid$globe_alt(
  id = 'my_globe_alt_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the globe_alt icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/globe_alt.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/globe_alt.svg}
}
\keyword{globe_alt}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
