% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{hand}
\alias{hand}
\title{hand}
\usage{
outline$hand()

solid$hand()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''hand'

Returns the svg markup for the heroicon ''hand'
}
\description{
hand

hand
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/hand.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/hand.svg}
}
\keyword{hand}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
