% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{hashtag}
\alias{hashtag}
\title{hashtag}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''hashtag'

Returns the svg markup for the heroicon ''hashtag'
}
\description{
hashtag

hashtag
}
\examples{
rheroicons::outline$hashtag(
  id = 'my_hashtag_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the hashtag icon'
)
rheroicons::solid$hashtag(
  id = 'my_hashtag_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the hashtag icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/hashtag.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/hashtag.svg}
}
\keyword{hashtag}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
