% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{home}
\alias{home}
\title{home}
\usage{
outline$home()

solid$home()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a home

create a solid or outline SVG icon of a home
}
\description{
home

home
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/home.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/home.svg}
\url{https://heroicons.dev}
}
\keyword{home}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
