% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{inbox}
\alias{inbox}
\title{inbox}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''inbox'

Returns the svg markup for the heroicon ''inbox'
}
\description{
inbox

inbox
}
\examples{
rheroicons::outline$inbox(
  id = 'my_inbox_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the inbox icon'
)
rheroicons::solid$inbox(
  id = 'my_inbox_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the inbox icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/inbox.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/inbox.svg}
}
\keyword{inbox}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
