% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{inbox}
\alias{inbox}
\title{inbox}
\usage{
outline$inbox()

solid$inbox()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a inbox

create a solid or outline SVG icon of a inbox
}
\description{
inbox

inbox
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/inbox.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/inbox.svg}
\url{https://heroicons.dev}
}
\keyword{inbox}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
