% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{inbox_in}
\alias{inbox_in}
\title{inbox_in}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''inbox_in'

Returns the svg markup for the heroicon ''inbox_in'
}
\description{
inbox_in

inbox_in
}
\examples{
rheroicons::outline$inbox_in(
  id = 'my_inbox_in_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the inbox_in icon'
)
rheroicons::solid$inbox_in(
  id = 'my_inbox_in_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the inbox_in icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/inbox_in.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/inbox_in.svg}
}
\keyword{inbox_in}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
