% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{inbox_in}
\alias{inbox_in}
\title{inbox_in}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a inbox_in
}
\description{
Render an svg icon of a inbox_in
}
\examples{
rheroicons::icons$inbox_in(
  type = "solid",
  id = "my_inbox_in_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the inbox_in icon"
)
}
\keyword{inbox_in}
\keyword{rheroicons}
