% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{information_circle}
\alias{information_circle}
\title{information_circle}
\usage{
outline$information_circle()

solid$information_circle()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a information_circle

create a solid or outline SVG icon of a information_circle
}
\description{
information_circle

information_circle
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/information_circle.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/information_circle.svg}
\url{https://heroicons.dev}
}
\keyword{information_circle}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
