% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{key}
\alias{key}
\title{key}
\usage{
outline$key()

solid$key()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''key'

Returns the svg markup for the heroicon ''key'
}
\description{
key

key
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/key.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/key.svg}
}
\keyword{key}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
