% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_gallery.R
\name{launch_gallery}
\alias{launch_gallery}
\title{\code{launch_gallery}}
\usage{
launch_gallery()
}
\value{
This function starts the rheroicons gallery
}
\description{
This function starts the rheroicons gallery
}
\examples{
\dontrun{
  rheroicons::launch_gallery()
}

}
\keyword{demo}
\keyword{gallery}
\keyword{rheroicons}
