% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_gallery.R
\name{launch_gallery}
\alias{launch_gallery}
\title{Icon gallery}
\usage{
launch_gallery(...)
}
\arguments{
\item{...}{additional 'shiny' options}
}
\value{
A 'shiny' app containing the icon gallery
}
\description{
Run the gallery to view and search for icons. Click on an icon
name to copy the 'R' code to the clipboard.
}
\examples{
if (interactive()) {
  rheroicons::launch_gallery()
}

}
