% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{library}
\alias{library}
\title{library}
\usage{
outline$library()

solid$library()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''library'

Returns the svg markup for the heroicon ''library'
}
\description{
library

library
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/library.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/library.svg}
}
\keyword{library}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
