% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{light_bulb}
\alias{light_bulb}
\title{light_bulb}
\usage{
outline$light_bulb()

solid$light_bulb()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''light_bulb'

Returns the svg markup for the heroicon ''light_bulb'
}
\description{
light_bulb

light_bulb
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/light_bulb.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/light_bulb.svg}
}
\keyword{light_bulb}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
