% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{light_bulb}
\alias{light_bulb}
\title{light_bulb}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a light_bulb
}
\description{
Render an svg icon of a light_bulb
}
\examples{
rheroicons::icons$light_bulb(
  type = "solid",
  id = "my_light_bulb_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the light_bulb icon"
)
}
\keyword{light_bulb}
\keyword{rheroicons}
