% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{lightning_bolt}
\alias{lightning_bolt}
\title{lightning_bolt}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''lightning_bolt'

Returns the svg markup for the heroicon ''lightning_bolt'
}
\description{
lightning_bolt

lightning_bolt
}
\examples{
rheroicons::outline$lightning_bolt(
  id = 'my_lightning_bolt_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the lightning_bolt icon'
)
rheroicons::solid$lightning_bolt(
  id = 'my_lightning_bolt_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the lightning_bolt icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/lightning_bolt.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/lightning_bolt.svg}
}
\keyword{lightning_bolt}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
