% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{lightning_bolt}
\alias{lightning_bolt}
\title{lightning_bolt}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a lightning_bolt
}
\description{
Render an svg icon of a lightning_bolt
}
\examples{
rheroicons::icons$lightning_bolt(
  type = "solid",
  id = "my_lightning_bolt_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the lightning_bolt icon"
)
}
\keyword{lightning_bolt}
\keyword{rheroicons}
