% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{link}
\alias{link}
\title{link}
\usage{
outline$link()

solid$link()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a link

create a solid or outline SVG icon of a link
}
\description{
link

link
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/link.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/link.svg}
\url{https://heroicons.dev}
}
\keyword{link}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
