% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{location_marker}
\alias{location_marker}
\title{location_marker}
\usage{
outline$location_marker()

solid$location_marker()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a location_marker

create a solid or outline SVG icon of a location_marker
}
\description{
location_marker

location_marker
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/location_marker.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/location_marker.svg}
\url{https://heroicons.dev}
}
\keyword{location_marker}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
