% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{lock_closed}
\alias{lock_closed}
\title{lock_closed}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a lock_closed
}
\description{
Render an svg icon of a lock_closed
}
\examples{
rheroicons::icons$lock_closed(
  type = "solid",
  id = "my_lock_closed_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the lock_closed icon"
)
}
\keyword{lock_closed}
\keyword{rheroicons}
