% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{lock_closed}
\alias{lock_closed}
\title{lock_closed}
\usage{
outline$lock_closed()

solid$lock_closed()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a lock_closed

create a solid or outline SVG icon of a lock_closed
}
\description{
lock_closed

lock_closed
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/lock_closed.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/lock_closed.svg}
\url{https://heroicons.dev}
}
\keyword{lock_closed}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
