% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{lock_open}
\alias{lock_open}
\title{lock_open}
\usage{
outline$lock_open()

solid$lock_open()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a lock_open

create a solid or outline SVG icon of a lock_open
}
\description{
lock_open

lock_open
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/lock_open.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/lock_open.svg}
\url{https://heroicons.dev}
}
\keyword{lock_open}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
