% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{logout}
\alias{logout}
\title{logout}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''logout'

Returns the svg markup for the heroicon ''logout'
}
\description{
logout

logout
}
\examples{
rheroicons::outline$logout(
  id = 'my_logout_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the logout icon'
)
rheroicons::solid$logout(
  id = 'my_logout_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the logout icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/logout.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/logout.svg}
}
\keyword{logout}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
