% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{mail}
\alias{mail}
\title{mail}
\usage{
outline$mail()

solid$mail()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a mail

create a solid or outline SVG icon of a mail
}
\description{
mail

mail
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/mail.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/mail.svg}
\url{https://heroicons.dev}
}
\keyword{mail}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
