% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{menu}
\alias{menu}
\title{menu}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''menu'

Returns the svg markup for the heroicon ''menu'
}
\description{
menu

menu
}
\examples{
rheroicons::outline$menu(
  id = 'my_menu_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the menu icon'
)
rheroicons::solid$menu(
  id = 'my_menu_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the menu icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/menu.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/menu.svg}
}
\keyword{menu}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
