% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{menu_alt_2}
\alias{menu_alt_2}
\title{menu_alt_2}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a menu_alt_2
}
\description{
Render an svg icon of a menu_alt_2
}
\examples{
rheroicons::icons$menu_alt_2(
  type = "solid",
  id = "my_menu_alt_2_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the menu_alt_2 icon"
)
}
\keyword{menu_alt_2}
\keyword{rheroicons}
