% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{menu_alt_3}
\alias{menu_alt_3}
\title{menu_alt_3}
\usage{
outline$menu_alt_3()

solid$menu_alt_3()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a menu_alt_3

create a solid or outline SVG icon of a menu_alt_3
}
\description{
menu_alt_3

menu_alt_3
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/menu_alt_3.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/menu_alt_3.svg}
\url{https://heroicons.dev}
}
\keyword{menu_alt_3}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
