% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{menu_alt_4}
\alias{menu_alt_4}
\title{menu_alt_4}
\usage{
outline$menu_alt_4()

solid$menu_alt_4()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''menu_alt_4'

Returns the svg markup for the heroicon ''menu_alt_4'
}
\description{
menu_alt_4

menu_alt_4
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/menu_alt_4.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/menu_alt_4.svg}
}
\keyword{menu_alt_4}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
