% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{microphone}
\alias{microphone}
\title{microphone}
\usage{
outline$microphone()

solid$microphone()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a microphone

create a solid or outline SVG icon of a microphone
}
\description{
microphone

microphone
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/microphone.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/microphone.svg}
\url{https://heroicons.dev}
}
\keyword{microphone}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
