% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{microphone}
\alias{microphone}
\title{microphone}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''microphone'

Returns the svg markup for the heroicon ''microphone'
}
\description{
microphone

microphone
}
\examples{
rheroicons::outline$microphone(
  id = 'my_microphone_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the microphone icon'
)
rheroicons::solid$microphone(
  id = 'my_microphone_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the microphone icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/microphone.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/microphone.svg}
}
\keyword{microphone}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
