% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{minus_circle}
\alias{minus_circle}
\title{minus_circle}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''minus_circle'

Returns the svg markup for the heroicon ''minus_circle'
}
\description{
minus_circle

minus_circle
}
\examples{
rheroicons::outline$minus_circle(
  id = 'my_minus_circle_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the minus_circle icon'
)
rheroicons::solid$minus_circle(
  id = 'my_minus_circle_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the minus_circle icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/minus_circle.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/minus_circle.svg}
}
\keyword{minus_circle}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
