% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{minus_circle}
\alias{minus_circle}
\title{minus_circle}
\usage{
outline$minus_circle()

solid$minus_circle()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a minus_circle

create a solid or outline SVG icon of a minus_circle
}
\description{
minus_circle

minus_circle
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/minus_circle.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/minus_circle.svg}
\url{https://heroicons.dev}
}
\keyword{minus_circle}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
