% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{moon}
\alias{moon}
\title{moon}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''moon'

Returns the svg markup for the heroicon ''moon'
}
\description{
moon

moon
}
\examples{
rheroicons::outline$moon(
  id = 'my_moon_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the moon icon'
)
rheroicons::solid$moon(
  id = 'my_moon_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the moon icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/moon.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/moon.svg}
}
\keyword{moon}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
