% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{newspaper}
\alias{newspaper}
\title{newspaper}
\usage{
outline$newspaper()

solid$newspaper()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a newspaper

create a solid or outline SVG icon of a newspaper
}
\description{
newspaper

newspaper
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/newspaper.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/newspaper.svg}
\url{https://heroicons.dev}
}
\keyword{newspaper}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
