% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R
\docType{data}
\name{outline}
\alias{outline}
\title{OUTLINE SVG Icons}
\format{
An object of class \code{list} of length 176.
}
\usage{
rheroicons::outline$icon_name()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers}
}
\description{
OUTLINE SVG Icons
}
\details{
outline icons from heroicons
}
\examples{
reheroicons::outline$book_open()
reheroicons::outline$book_open(id = 'myBookIcon')
reheroicons::outline$book_open(class = 'my-icon-set')
}
\references{
\url{https://github.com/refactoringui/heroicons}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{rheroicons}
