% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{paper_clip}
\alias{paper_clip}
\title{paper_clip}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''paper_clip'

Returns the svg markup for the heroicon ''paper_clip'
}
\description{
paper_clip

paper_clip
}
\examples{
rheroicons::outline$paper_clip(
  id = 'my_paper_clip_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the paper_clip icon'
)
rheroicons::solid$paper_clip(
  id = 'my_paper_clip_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the paper_clip icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/paper_clip.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/paper_clip.svg}
}
\keyword{outline}
\keyword{paper_clip}
\keyword{rheroicons}
\keyword{solid}
