% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{pause}
\alias{pause}
\title{pause}
\usage{
outline$pause()

solid$pause()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''pause'

Returns the svg markup for the heroicon ''pause'
}
\description{
pause

pause
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/pause.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/pause.svg}
}
\keyword{outline}
\keyword{pause}
\keyword{rheroicons}
\keyword{solid}
