% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{pause}
\alias{pause}
\title{pause}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''pause'

Returns the svg markup for the heroicon ''pause'
}
\description{
pause

pause
}
\examples{
rheroicons::outline$pause(
  id = 'my_pause_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the pause icon'
)
rheroicons::solid$pause(
  id = 'my_pause_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the pause icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/pause.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/pause.svg}
}
\keyword{outline}
\keyword{pause}
\keyword{rheroicons}
\keyword{solid}
