% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{pencil}
\alias{pencil}
\title{pencil}
\usage{
outline$pencil()

solid$pencil()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a pencil

create a solid or outline SVG icon of a pencil
}
\description{
pencil

pencil
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/pencil.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/pencil.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{pencil}
\keyword{rheroicons}
\keyword{solid}
