% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{pencil_alt}
\alias{pencil_alt}
\title{pencil_alt}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a pencil_alt
}
\description{
Render an svg icon of a pencil_alt
}
\examples{
rheroicons::icons$pencil_alt(
  type = "solid",
  id = "my_pencil_alt_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the pencil_alt icon"
)
}
\keyword{pencil_alt}
\keyword{rheroicons}
