% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{pencil_alt}
\alias{pencil_alt}
\title{pencil_alt}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''pencil_alt'

Returns the svg markup for the heroicon ''pencil_alt'
}
\description{
pencil_alt

pencil_alt
}
\examples{
rheroicons::outline$pencil_alt(
  id = 'my_pencil_alt_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the pencil_alt icon'
)
rheroicons::solid$pencil_alt(
  id = 'my_pencil_alt_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the pencil_alt icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/pencil_alt.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/pencil_alt.svg}
}
\keyword{outline}
\keyword{pencil_alt}
\keyword{rheroicons}
\keyword{solid}
