% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{phone_outgoing}
\alias{phone_outgoing}
\title{phone_outgoing}
\usage{
outline$phone_outgoing()

solid$phone_outgoing()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a phone_outgoing

create a solid or outline SVG icon of a phone_outgoing
}
\description{
phone_outgoing

phone_outgoing
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/phone_outgoing.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/phone_outgoing.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{phone_outgoing}
\keyword{rheroicons}
\keyword{solid}
