% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{phone_outgoing}
\alias{phone_outgoing}
\title{phone_outgoing}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''phone_outgoing'

Returns the svg markup for the heroicon ''phone_outgoing'
}
\description{
phone_outgoing

phone_outgoing
}
\examples{
rheroicons::outline$phone_outgoing(
  id = 'my_phone_outgoing_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the phone_outgoing icon'
)
rheroicons::solid$phone_outgoing(
  id = 'my_phone_outgoing_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the phone_outgoing icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/phone_outgoing.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/phone_outgoing.svg}
}
\keyword{outline}
\keyword{phone_outgoing}
\keyword{rheroicons}
\keyword{solid}
