% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{photograph}
\alias{photograph}
\title{photograph}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''photograph'

Returns the svg markup for the heroicon ''photograph'
}
\description{
photograph

photograph
}
\examples{
rheroicons::outline$photograph(
  id = 'my_photograph_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the photograph icon'
)
rheroicons::solid$photograph(
  id = 'my_photograph_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the photograph icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/photograph.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/photograph.svg}
}
\keyword{outline}
\keyword{photograph}
\keyword{rheroicons}
\keyword{solid}
