% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{photograph}
\alias{photograph}
\title{photograph}
\usage{
outline$photograph()

solid$photograph()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a photograph

create a solid or outline SVG icon of a photograph
}
\description{
photograph

photograph
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/photograph.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/photograph.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{photograph}
\keyword{rheroicons}
\keyword{solid}
