% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{play}
\alias{play}
\title{play}
\usage{
outline$play()

solid$play()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a play

create a solid or outline SVG icon of a play
}
\description{
play

play
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/play.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/play.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{play}
\keyword{rheroicons}
\keyword{solid}
