% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{play}
\alias{play}
\title{play}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''play'

Returns the svg markup for the heroicon ''play'
}
\description{
play

play
}
\examples{
rheroicons::outline$play(
  id = 'my_play_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the play icon'
)
rheroicons::solid$play(
  id = 'my_play_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the play icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/play.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/play.svg}
}
\keyword{outline}
\keyword{play}
\keyword{rheroicons}
\keyword{solid}
