% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{plus}
\alias{plus}
\title{plus}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''plus'

Returns the svg markup for the heroicon ''plus'
}
\description{
plus

plus
}
\examples{
rheroicons::outline$plus(
  id = 'my_plus_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the plus icon'
)
rheroicons::solid$plus(
  id = 'my_plus_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the plus icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/plus.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/plus.svg}
}
\keyword{outline}
\keyword{plus}
\keyword{rheroicons}
\keyword{solid}
