% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{plus}
\alias{plus}
\title{plus}
\usage{
outline$plus()

solid$plus()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a plus

create a solid or outline SVG icon of a plus
}
\description{
plus

plus
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/plus.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/plus.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{plus}
\keyword{rheroicons}
\keyword{solid}
