% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{plus_circle}
\alias{plus_circle}
\title{plus_circle}
\usage{
outline$plus_circle()

solid$plus_circle()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''plus_circle'

Returns the svg markup for the heroicon ''plus_circle'
}
\description{
plus_circle

plus_circle
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/plus_circle.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/plus_circle.svg}
}
\keyword{outline}
\keyword{plus_circle}
\keyword{rheroicons}
\keyword{solid}
