% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{printer}
\alias{printer}
\title{printer}
\usage{
outline$printer()

solid$printer()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''printer'

Returns the svg markup for the heroicon ''printer'
}
\description{
printer

printer
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/printer.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/printer.svg}
}
\keyword{outline}
\keyword{printer}
\keyword{rheroicons}
\keyword{solid}
