% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{puzzle}
\alias{puzzle}
\title{puzzle}
\usage{
outline$puzzle()

solid$puzzle()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a puzzle

create a solid or outline SVG icon of a puzzle
}
\description{
puzzle

puzzle
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/puzzle.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/puzzle.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{puzzle}
\keyword{rheroicons}
\keyword{solid}
