% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{qrcode}
\alias{qrcode}
\title{qrcode}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''qrcode'

Returns the svg markup for the heroicon ''qrcode'
}
\description{
qrcode

qrcode
}
\examples{
rheroicons::outline$qrcode(
  id = 'my_qrcode_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the qrcode icon'
)
rheroicons::solid$qrcode(
  id = 'my_qrcode_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the qrcode icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/qrcode.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/qrcode.svg}
}
\keyword{outline}
\keyword{qrcode}
\keyword{rheroicons}
\keyword{solid}
