% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{receipt_refund}
\alias{receipt_refund}
\title{receipt_refund}
\usage{
outline$receipt_refund()

solid$receipt_refund()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a receipt_refund

create a solid or outline SVG icon of a receipt_refund
}
\description{
receipt_refund

receipt_refund
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/receipt_refund.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/receipt_refund.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{receipt_refund}
\keyword{rheroicons}
\keyword{solid}
