% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{receipt_refund}
\alias{receipt_refund}
\title{receipt_refund}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a receipt_refund
}
\description{
Render an svg icon of a receipt_refund
}
\examples{
rheroicons::icons$receipt_refund(
  type = "solid",
  id = "my_receipt_refund_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the receipt_refund icon"
)
}
\keyword{receipt_refund}
\keyword{rheroicons}
