% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{refresh}
\alias{refresh}
\title{refresh}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''refresh'

Returns the svg markup for the heroicon ''refresh'
}
\description{
refresh

refresh
}
\examples{
rheroicons::outline$refresh(
  id = 'my_refresh_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the refresh icon'
)
rheroicons::solid$refresh(
  id = 'my_refresh_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the refresh icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/refresh.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/refresh.svg}
}
\keyword{outline}
\keyword{refresh}
\keyword{rheroicons}
\keyword{solid}
