% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{reply}
\alias{reply}
\title{reply}
\usage{
outline$reply()

solid$reply()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''reply'

Returns the svg markup for the heroicon ''reply'
}
\description{
reply

reply
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/reply.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/reply.svg}
}
\keyword{outline}
\keyword{reply}
\keyword{rheroicons}
\keyword{solid}
