% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{scale}
\alias{scale}
\title{scale}
\usage{
outline$scale()

solid$scale()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a scale

create a solid or outline SVG icon of a scale
}
\description{
scale

scale
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/scale.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/scale.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{scale}
\keyword{solid}
