% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{search}
\alias{search}
\title{search}
\usage{
outline$search()

solid$search()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''search'

Returns the svg markup for the heroicon ''search'
}
\description{
search

search
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/search.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/search.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{search}
\keyword{solid}
