% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{share}
\alias{share}
\title{share}
\usage{
outline$share()

solid$share()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''share'

Returns the svg markup for the heroicon ''share'
}
\description{
share

share
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/share.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/share.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{share}
\keyword{solid}
