% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{shield_check}
\alias{shield_check}
\title{shield_check}
\usage{
outline$shield_check()

solid$shield_check()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''shield_check'

Returns the svg markup for the heroicon ''shield_check'
}
\description{
shield_check

shield_check
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/shield_check.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/shield_check.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{shield_check}
\keyword{solid}
