% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{shopping_bag}
\alias{shopping_bag}
\title{shopping_bag}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''shopping_bag'

Returns the svg markup for the heroicon ''shopping_bag'
}
\description{
shopping_bag

shopping_bag
}
\examples{
rheroicons::outline$shopping_bag(
  id = 'my_shopping_bag_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the shopping_bag icon'
)
rheroicons::solid$shopping_bag(
  id = 'my_shopping_bag_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the shopping_bag icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/shopping_bag.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/shopping_bag.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{shopping_bag}
\keyword{solid}
