% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{shopping_cart}
\alias{shopping_cart}
\title{shopping_cart}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''shopping_cart'

Returns the svg markup for the heroicon ''shopping_cart'
}
\description{
shopping_cart

shopping_cart
}
\examples{
rheroicons::outline$shopping_cart(
  id = 'my_shopping_cart_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the shopping_cart icon'
)
rheroicons::solid$shopping_cart(
  id = 'my_shopping_cart_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the shopping_cart icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/shopping_cart.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/shopping_cart.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{shopping_cart}
\keyword{solid}
