% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{shopping_cart}
\alias{shopping_cart}
\title{shopping_cart}
\usage{
outline$shopping_cart()

solid$shopping_cart()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''shopping_cart'

Returns the svg markup for the heroicon ''shopping_cart'
}
\description{
shopping_cart

shopping_cart
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/shopping_cart.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/shopping_cart.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{shopping_cart}
\keyword{solid}
