% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solid.R
\docType{data}
\name{solid}
\alias{solid}
\title{SOLID SVG Icons}
\format{
An object of class \code{list} of length 176.
}
\usage{
rheroicons::solid$icon_name()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers}
}
\value{
solid heroicons
}
\description{
SOLID SVG Icons
}
\examples{
reheroicons::solid$book_open()
reheroicons::solid$book_open(id = 'myBookIcon')
reheroicons::solid$book_open(class = 'my-icon-set')
}
\references{
\url{https://github.com/refactoringui/heroicons}
\url{https://heroicons.dev}
}
\keyword{rheroicons}
\keyword{solid}
