% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{sort_ascending}
\alias{sort_ascending}
\title{sort_ascending}
\usage{
outline$sort_ascending()

solid$sort_ascending()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a sort_ascending

create a solid or outline SVG icon of a sort_ascending
}
\description{
sort_ascending

sort_ascending
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/sort_ascending.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/sort_ascending.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{sort_ascending}
