% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{sort_descending}
\alias{sort_descending}
\title{sort_descending}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a sort_descending
}
\description{
Render an svg icon of a sort_descending
}
\examples{
rheroicons::icons$sort_descending(
  type = "solid",
  id = "my_sort_descending_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the sort_descending icon"
)
}
\keyword{rheroicons}
\keyword{sort_descending}
