% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{sort_descending}
\alias{sort_descending}
\title{sort_descending}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''sort_descending'

Returns the svg markup for the heroicon ''sort_descending'
}
\description{
sort_descending

sort_descending
}
\examples{
rheroicons::outline$sort_descending(
  id = 'my_sort_descending_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the sort_descending icon'
)
rheroicons::solid$sort_descending(
  id = 'my_sort_descending_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the sort_descending icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/sort_descending.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/sort_descending.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{sort_descending}
