% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{speakerphone}
\alias{speakerphone}
\title{speakerphone}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''speakerphone'

Returns the svg markup for the heroicon ''speakerphone'
}
\description{
speakerphone

speakerphone
}
\examples{
rheroicons::outline$speakerphone(
  id = 'my_speakerphone_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the speakerphone icon'
)
rheroicons::solid$speakerphone(
  id = 'my_speakerphone_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the speakerphone icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/speakerphone.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/speakerphone.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{speakerphone}
