% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{speakerphone}
\alias{speakerphone}
\title{speakerphone}
\usage{
outline$speakerphone()

solid$speakerphone()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a speakerphone

create a solid or outline SVG icon of a speakerphone
}
\description{
speakerphone

speakerphone
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/speakerphone.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/speakerphone.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{speakerphone}
