% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{stop}
\alias{stop}
\title{stop}
\usage{
outline$stop()

solid$stop()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''stop'

Returns the svg markup for the heroicon ''stop'
}
\description{
stop

stop
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/stop.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/stop.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{stop}
