% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{sun}
\alias{sun}
\title{sun}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''sun'

Returns the svg markup for the heroicon ''sun'
}
\description{
sun

sun
}
\examples{
rheroicons::outline$sun(
  id = 'my_sun_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the sun icon'
)
rheroicons::solid$sun(
  id = 'my_sun_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the sun icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/sun.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/sun.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{sun}
