% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{switch_horizontal}
\alias{switch_horizontal}
\title{switch_horizontal}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''switch_horizontal'

Returns the svg markup for the heroicon ''switch_horizontal'
}
\description{
switch_horizontal

switch_horizontal
}
\examples{
rheroicons::outline$switch_horizontal(
  id = 'my_switch_horizontal_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the switch_horizontal icon'
)
rheroicons::solid$switch_horizontal(
  id = 'my_switch_horizontal_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the switch_horizontal icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/switch_horizontal.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/switch_horizontal.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{switch_horizontal}
