% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{switch_horizontal}
\alias{switch_horizontal}
\title{switch_horizontal}
\usage{
outline$switch_horizontal()

solid$switch_horizontal()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a switch_horizontal

create a solid or outline SVG icon of a switch_horizontal
}
\description{
switch_horizontal

switch_horizontal
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/switch_horizontal.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/switch_horizontal.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{switch_horizontal}
