% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{switch_vertical}
\alias{switch_vertical}
\title{switch_vertical}
\usage{
outline$switch_vertical()

solid$switch_vertical()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''switch_vertical'

Returns the svg markup for the heroicon ''switch_vertical'
}
\description{
switch_vertical

switch_vertical
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/switch_vertical.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/switch_vertical.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{switch_vertical}
