% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{tag}
\alias{tag}
\title{tag}
\usage{
outline$tag()

solid$tag()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''tag'

Returns the svg markup for the heroicon ''tag'
}
\description{
tag

tag
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/tag.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/tag.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{tag}
