% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{template}
\alias{template}
\title{template}
\usage{
outline$template()

solid$template()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''template'

Returns the svg markup for the heroicon ''template'
}
\description{
template

template
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/template.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/template.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{template}
