% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{template}
\alias{template}
\title{template}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''template'

Returns the svg markup for the heroicon ''template'
}
\description{
template

template
}
\examples{
rheroicons::outline$template(
  id = 'my_template_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the template icon'
)
rheroicons::solid$template(
  id = 'my_template_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the template icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/template.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/template.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{template}
