% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{terminal}
\alias{terminal}
\title{terminal}
\usage{
outline$terminal()

solid$terminal()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a terminal

create a solid or outline SVG icon of a terminal
}
\description{
terminal

terminal
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/terminal.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/terminal.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{terminal}
