% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{thumb_down}
\alias{thumb_down}
\title{thumb_down}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''thumb_down'

Returns the svg markup for the heroicon ''thumb_down'
}
\description{
thumb_down

thumb_down
}
\examples{
rheroicons::outline$thumb_down(
  id = 'my_thumb_down_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the thumb_down icon'
)
rheroicons::solid$thumb_down(
  id = 'my_thumb_down_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the thumb_down icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/thumb_down.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/thumb_down.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{thumb_down}
