% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{thumb_up}
\alias{thumb_up}
\title{thumb_up}
\usage{
outline$thumb_up()

solid$thumb_up()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''thumb_up'

Returns the svg markup for the heroicon ''thumb_up'
}
\description{
thumb_up

thumb_up
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/thumb_up.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/thumb_up.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{thumb_up}
