% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{ticket}
\alias{ticket}
\title{ticket}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''ticket'

Returns the svg markup for the heroicon ''ticket'
}
\description{
ticket

ticket
}
\examples{
rheroicons::outline$ticket(
  id = 'my_ticket_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the ticket icon'
)
rheroicons::solid$ticket(
  id = 'my_ticket_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the ticket icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/ticket.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/ticket.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{ticket}
