% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{translate}
\alias{translate}
\title{translate}
\usage{
outline$translate()

solid$translate()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''translate'

Returns the svg markup for the heroicon ''translate'
}
\description{
translate

translate
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/translate.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/translate.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{translate}
