% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{trash}
\alias{trash}
\title{trash}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''trash'

Returns the svg markup for the heroicon ''trash'
}
\description{
trash

trash
}
\examples{
rheroicons::outline$trash(
  id = 'my_trash_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the trash icon'
)
rheroicons::solid$trash(
  id = 'my_trash_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the trash icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/trash.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/trash.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{trash}
