% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{trash}
\alias{trash}
\title{trash}
\usage{
outline$trash()

solid$trash()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''trash'

Returns the svg markup for the heroicon ''trash'
}
\description{
trash

trash
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/trash.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/trash.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{trash}
