% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{trending_down}
\alias{trending_down}
\title{trending_down}
\usage{
outline$trending_down()

solid$trending_down()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''trending_down'

Returns the svg markup for the heroicon ''trending_down'
}
\description{
trending_down

trending_down
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/trending_down.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/trending_down.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{trending_down}
