% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{upload}
\alias{upload}
\title{upload}
\usage{
outline$upload()

solid$upload()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a upload

create a solid or outline SVG icon of a upload
}
\description{
upload

upload
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/upload.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/upload.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{upload}
