% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{user_add}
\alias{user_add}
\title{user_add}
\usage{
outline$user_add()

solid$user_add()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a user_add

create a solid or outline SVG icon of a user_add
}
\description{
user_add

user_add
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/user_add.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/user_add.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{user_add}
