% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{user_circle}
\alias{user_circle}
\title{user_circle}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''user_circle'

Returns the svg markup for the heroicon ''user_circle'
}
\description{
user_circle

user_circle
}
\examples{
rheroicons::outline$user_circle(
  id = 'my_user_circle_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the user_circle icon'
)
rheroicons::solid$user_circle(
  id = 'my_user_circle_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the user_circle icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/user_circle.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/user_circle.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{user_circle}
