% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{user_group}
\alias{user_group}
\title{user_group}
\usage{
outline$user_group()

solid$user_group()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a user_group

create a solid or outline SVG icon of a user_group
}
\description{
user_group

user_group
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/user_group.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/user_group.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{user_group}
