% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{user_remove}
\alias{user_remove}
\title{user_remove}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''user_remove'

Returns the svg markup for the heroicon ''user_remove'
}
\description{
user_remove

user_remove
}
\examples{
rheroicons::outline$user_remove(
  id = 'my_user_remove_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the user_remove icon'
)
rheroicons::solid$user_remove(
  id = 'my_user_remove_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the user_remove icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/user_remove.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/user_remove.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{user_remove}
