% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{user_remove}
\alias{user_remove}
\title{user_remove}
\usage{
outline$user_remove()

solid$user_remove()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a user_remove

create a solid or outline SVG icon of a user_remove
}
\description{
user_remove

user_remove
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/user_remove.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/user_remove.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{user_remove}
