% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{users}
\alias{users}
\title{users}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''users'

Returns the svg markup for the heroicon ''users'
}
\description{
users

users
}
\examples{
rheroicons::outline$users(
  id = 'my_users_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the users icon'
)
rheroicons::solid$users(
  id = 'my_users_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the users icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/users.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/users.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{users}
