% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{view_boards}
\alias{view_boards}
\title{view_boards}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''view_boards'

Returns the svg markup for the heroicon ''view_boards'
}
\description{
view_boards

view_boards
}
\examples{
rheroicons::outline$view_boards(
  id = 'my_view_boards_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the view_boards icon'
)
rheroicons::solid$view_boards(
  id = 'my_view_boards_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the view_boards icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/view_boards.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/view_boards.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{view_boards}
