% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{view_grid_add}
\alias{view_grid_add}
\title{view_grid_add}
\usage{
outline$view_grid_add()

solid$view_grid_add()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a view_grid_add

create a solid or outline SVG icon of a view_grid_add
}
\description{
view_grid_add

view_grid_add
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/view_grid_add.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/view_grid_add.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{view_grid_add}
