% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{view_list}
\alias{view_list}
\title{view_list}
\usage{
outline$view_list()

solid$view_list()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''view_list'

Returns the svg markup for the heroicon ''view_list'
}
\description{
view_list

view_list
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/view_list.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/view_list.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{view_list}
