% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{volume_off}
\alias{volume_off}
\title{volume_off}
\usage{
outline$volume_off()

solid$volume_off()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a volume_off

create a solid or outline SVG icon of a volume_off
}
\description{
volume_off

volume_off
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/volume_off.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/volume_off.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{volume_off}
