% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{volume_up}
\alias{volume_up}
\title{volume_up}
\usage{
outline$volume_up()

solid$volume_up()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a volume_up

create a solid or outline SVG icon of a volume_up
}
\description{
volume_up

volume_up
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/volume_up.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/volume_up.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{volume_up}
