% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{x}
\alias{x}
\title{x}
\usage{
outline$x()

solid$x()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''x'

Returns the svg markup for the heroicon ''x'
}
\description{
x

x
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/x.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/x.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{x}
