% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{x}
\alias{x}
\title{x}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}

\item{title}{a string that describes the icon(should be used if aria_hidden is FALSE)}
}
\value{
Returns the svg markup for the heroicon ''x'

Returns the svg markup for the heroicon ''x'
}
\description{
x

x
}
\examples{
rheroicons::outline$x(
  id = 'my_x_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the x icon'
)
rheroicons::solid$x(
  id = 'my_x_icon',
  class = 'my-icons',
  aria_hidden = FALSE,
  title = 'a title for the x icon'
)
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/x.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/x.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{x}
