% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{x_circle}
\alias{x_circle}
\title{x_circle}
\usage{
outline$x_circle()

solid$x_circle()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
create a solid or outline SVG icon of a x_circle

create a solid or outline SVG icon of a x_circle
}
\description{
x_circle

x_circle
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/x_circle.svg}
\url{https://heroicons.dev}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/x_circle.svg}
\url{https://heroicons.dev}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{x_circle}
