% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R, R/solid.R
\name{zoom_in}
\alias{zoom_in}
\title{zoom_in}
\usage{
outline$zoom_in()

solid$zoom_in()
}
\arguments{
\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be readable by screen readers (default: false)}
}
\value{
Returns the svg markup for the heroicon ''zoom_in'

Returns the svg markup for the heroicon ''zoom_in'
}
\description{
zoom_in

zoom_in
}
\references{
\url{https://github.com/refactoringui/heroicons/blob/master/outline/zoom_in.svg}

\url{https://github.com/refactoringui/heroicons/blob/master/solid/zoom_in.svg}
}
\keyword{outline}
\keyword{rheroicons}
\keyword{solid}
\keyword{zoom_in}
