% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{zoom_in}
\alias{zoom_in}
\title{zoom_in}
\arguments{
\item{type}{render an outline or solid icon (default outline)}

\item{id}{a unique ID to be applied to the svg icon}

\item{class}{a css class to be applied to the svg icon}

\item{aria_hidden}{should the icon be hidden from screen readers (default: TRUE)}

\item{title}{a string that describes the icon (optional)}
}
\value{
Render an svg icon of a zoom_in
}
\description{
Render an svg icon of a zoom_in
}
\examples{
rheroicons::icons$zoom_in(
  type = "solid",
  id = "my_zoom_in_icon",
  class = "my-icons",
  aria_hidden = FALSE,
  title = "a title for the zoom_in icon"
)
}
\keyword{rheroicons}
\keyword{zoom_in}
